+++ app/components/search/GeoToggle.client.tsx
@@ -67,7 +67,7 @@ const GeoToggle = () => {
           <Switch
             checked={enabled}
             onChange={toggleSearch}
-            className="group relative self-center xl:self-start flex h-7 w-14 cursor-pointer rounded-full bg-activeIsland/20 p-1 transition-colors duration-200 ease-in-out focus:outline-none data-[focus]:outline-1 data-[focus]:outline-white data-[checked]:bg-activeIsland/50"
+            className="group relative self-center xl:self-start flex h-7 w-14 cursor-pointer rounded-full bg-island/20 p-1 transition-colors duration-200 ease-in-out focus:outline-none data-[focus]:outline-1 data-[focus]:outline-white data-[checked]:bg-island/50"
           >
             <span
               aria-hidden="true"
