--- app/mapStyles/counties.ts
@@ -1,3058 +1,0 @@
-import type { FeatureCollection } from "geojson";
-import type { StyleSpecification } from "maplibre-gl";
-
-export const simpleCounties: FeatureCollection = {
-  type: "FeatureCollection",
-  features: [
-    {
-      type: "Feature",
-      properties: {
-        id: "InlandCounties.1",
-        STATE: "GA",
-        CWA: "CHS",
-        COUNTYNAME: "Liberty",
-        FIPS: "13179",
-        TIME_ZONE: "E",
-        FE_AREA: "se",
-        LON: -81.4998,
-        LAT: 31.8308,
-        uuid: "2f0676be-462b-509e-8bb2-1cf287840f6f",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-81.7141966, 32.09221216],
-              [-81.7120966, 32.09111216],
-              [-81.7093966, 32.09131216],
-              [-81.7058966, 32.08891216],
-              [-81.7019966, 32.08951216],
-              [-81.6989966, 32.09071216],
-              [-81.6964966, 32.08831216],
-              [-81.6935966, 32.08801216],
-              [-81.6904966, 32.08671216],
-              [-81.6890966, 32.08391216],
-              [-81.6878966, 32.07991216],
-              [-81.6858966, 32.07761216],
-              [-81.6840966, 32.07621216],
-              [-81.6793966, 32.07571216],
-              [-81.6760966, 32.07371216],
-              [-81.6732966, 32.06861216],
-              [-81.6728966, 32.06471216],
-              [-81.6715966, 32.06371216],
-              [-81.6685966, 32.06301216],
-              [-81.6653966, 32.06071216],
-              [-81.6625966, 32.05841216],
-              [-81.6597966, 32.05781216],
-              [-81.6566966, 32.05871216],
-              [-81.6542966, 32.05841216],
-              [-81.6516966, 32.05831216],
-              [-81.6472966, 32.05341216],
-              [-81.6440966, 32.05331216],
-              [-81.6444966, 32.05131216],
-              [-81.6421966, 32.05081216],
-              [-81.6401966, 32.04821216],
-              [-81.6388966, 32.04551216],
-              [-81.6364966, 32.04381216],
-              [-81.6356966, 32.04121216],
-              [-81.6326966, 32.04001216],
-              [-81.6302966, 32.03791216],
-              [-81.6251966, 32.03501216],
-              [-81.6234966, 32.03261216],
-              [-81.6209966, 32.03151216],
-              [-81.6180966, 32.02881216],
-              [-81.6166966, 32.02131216],
-              [-81.6143966, 32.02001216],
-              [-81.6123966, 32.01781216],
-              [-81.6099966, 32.01641216],
-              [-81.6070966, 32.01341216],
-              [-81.6043966, 32.01111216],
-              [-81.6038966, 32.00901216],
-              [-81.6050966, 32.00691216],
-              [-81.6048966, 32.00371216],
-              [-81.6019966, 32.00151216],
-              [-81.6004966, 31.99801216],
-              [-81.5987966, 31.99561216],
-              [-81.6012966, 31.99411216],
-              [-81.5937966, 31.99031216],
-              [-81.5943966, 31.98681216],
-              [-81.5930966, 31.98351216],
-              [-81.5923966, 31.97981216],
-              [-81.5912966, 31.97701216],
-              [-81.5921966, 31.97451216],
-              [-81.5916966, 31.97161216],
-              [-81.5879966, 31.97021216],
-              [-81.5854966, 31.97151216],
-              [-81.5816966, 31.97121216],
-              [-81.5775966, 31.97221216],
-              [-81.5743966, 31.97091216],
-              [-81.5702966, 31.96771216],
-              [-81.5668966, 31.96571216],
-              [-81.5653966, 31.96321216],
-              [-81.5648966, 31.96061216],
-              [-81.5632966, 31.95951216],
-              [-81.5592966, 31.95931216],
-              [-81.5559966, 31.95891216],
-              [-81.5517966, 31.95791216],
-              [-81.5484966, 31.95991216],
-              [-81.5477966, 31.95791216],
-              [-81.5460966, 31.95891216],
-              [-81.5379966, 31.96031216],
-              [-81.5351966, 31.96051216],
-              [-81.5366966, 31.96281216],
-              [-81.5342966, 31.96341216],
-              [-81.5313966, 31.96301216],
-              [-81.5277966, 31.96191216],
-              [-81.5232966, 31.96011216],
-              [-81.5199966, 31.95951216],
-              [-81.5170966, 31.96051216],
-              [-81.5147966, 31.96241216],
-              [-81.5156966, 31.96581216],
-              [-81.5131966, 31.96821216],
-              [-81.5100966, 31.96871216],
-              [-81.5085966, 31.96641216],
-              [-81.5061966, 31.96461216],
-              [-81.5027966, 31.96301216],
-              [-81.5002966, 31.96491216],
-              [-81.4983966, 31.96661216],
-              [-81.4953966, 31.96841216],
-              [-81.4906966, 31.96771216],
-              [-81.4886966, 31.96681216],
-              [-81.4862966, 31.96511216],
-              [-81.4832966, 31.96611216],
-              [-81.4845966, 31.96851216],
-              [-81.4826966, 31.97131216],
-              [-81.4795966, 31.97101216],
-              [-81.4805966, 31.97311216],
-              [-81.4776966, 31.97411216],
-              [-81.4745966, 31.97391216],
-              [-81.4716966, 31.97241216],
-              [-81.4695966, 31.96691216],
-              [-81.4637966, 31.95011216],
-              [-81.4273966, 31.93901216],
-              [-81.4239966, 31.93831216],
-              [-81.4212966, 31.93861216],
-              [-81.4173966, 31.93971216],
-              [-81.4156966, 31.94071216],
-              [-81.4138966, 31.94221216],
-              [-81.4116966, 31.94271216],
-              [-81.4077966, 31.94401216],
-              [-81.4062966, 31.94181216],
-              [-81.4052966, 31.93971216],
-              [-81.4047966, 31.93761216],
-              [-81.4036966, 31.93501216],
-              [-81.4001966, 31.92901216],
-              [-81.4013966, 31.92681216],
-              [-81.4009966, 31.92411216],
-              [-81.4017966, 31.92211216],
-              [-81.4040966, 31.92001216],
-              [-81.4028966, 31.91661216],
-              [-81.4005966, 31.91381216],
-              [-81.4005966, 31.91121216],
-              [-81.3985966, 31.90841216],
-              [-81.3961966, 31.90671216],
-              [-81.3945966, 31.90371216],
-              [-81.3947966, 31.90091216],
-              [-81.3932966, 31.89921216],
-              [-81.3921966, 31.89651216],
-              [-81.3917966, 31.89341216],
-              [-81.3933966, 31.89051216],
-              [-81.3950966, 31.88641216],
-              [-81.3958966, 31.88311216],
-              [-81.3930966, 31.88071216],
-              [-81.3906966, 31.87991216],
-              [-81.3874966, 31.87971216],
-              [-81.3848966, 31.87831216],
-              [-81.3829966, 31.87571216],
-              [-81.3830966, 31.87331216],
-              [-81.3821966, 31.87121216],
-              [-81.3809966, 31.86921216],
-              [-81.3810966, 31.86641216],
-              [-81.3828966, 31.86281216],
-              [-81.3828966, 31.85881216],
-              [-81.3817966, 31.85631216],
-              [-81.3814966, 31.85421216],
-              [-81.3787966, 31.85351216],
-              [-81.3774966, 31.84981216],
-              [-81.3732966, 31.84631216],
-              [-81.3704966, 31.84111216],
-              [-81.3692966, 31.84471216],
-              [-81.3673966, 31.84641216],
-              [-81.3649966, 31.84571216],
-              [-81.3636966, 31.84331216],
-              [-81.3641966, 31.84041216],
-              [-81.3625966, 31.83761216],
-              [-81.3601966, 31.83721216],
-              [-81.3576966, 31.83801216],
-              [-81.3545966, 31.83861216],
-              [-81.3532966, 31.83621216],
-              [-81.3565966, 31.83461216],
-              [-81.3580966, 31.83181216],
-              [-81.3550966, 31.82971216],
-              [-81.3519966, 31.82641216],
-              [-81.3474966, 31.82571216],
-              [-81.3449966, 31.82471216],
-              [-81.3453966, 31.82101216],
-              [-81.3475966, 31.81821216],
-              [-81.3517966, 31.81951216],
-              [-81.3552966, 31.81901216],
-              [-81.3555966, 31.81521216],
-              [-81.3520966, 31.81451216],
-              [-81.3485966, 31.81441216],
-              [-81.3457966, 31.81501216],
-              [-81.3417966, 31.81671216],
-              [-81.3392966, 31.81991216],
-              [-81.3425966, 31.82701216],
-              [-81.3415966, 31.82951216],
-              [-81.3391966, 31.83061216],
-              [-81.3360966, 31.83101216],
-              [-81.3331966, 31.83231216],
-              [-81.3333966, 31.83521216],
-              [-81.3371966, 31.83561216],
-              [-81.3406966, 31.83461216],
-              [-81.3432966, 31.83431216],
-              [-81.3452966, 31.83611216],
-              [-81.3432966, 31.83901216],
-              [-81.3392966, 31.84061216],
-              [-81.3358966, 31.84081216],
-              [-81.3334966, 31.84051216],
-              [-81.3307966, 31.83921216],
-              [-81.3289966, 31.83711216],
-              [-81.3279966, 31.83401216],
-              [-81.3263966, 31.83041216],
-              [-81.3271966, 31.82781216],
-              [-81.3285966, 31.82531216],
-              [-81.3277966, 31.82291216],
-              [-81.3243966, 31.82201216],
-              [-81.3202966, 31.82021216],
-              [-81.3167966, 31.81841216],
-              [-81.3155966, 31.81561216],
-              [-81.3172966, 31.81261216],
-              [-81.3192966, 31.81081216],
-              [-81.3196966, 31.80891216],
-              [-81.3190966, 31.80631216],
-              [-81.31679535, 31.80391312],
-              [-81.3144989, 31.80261231],
-              [-81.31349945, 31.80101204],
-              [-81.31089783, 31.79851151],
-              [-81.3102951, 31.79511261],
-              [-81.31139374, 31.7924118],
-              [-81.31090976957465, 31.789808829437057],
-              [-81.30649567, 31.78861237],
-              [-81.30329895, 31.78861237],
-              [-81.29819488, 31.78931236],
-              [-81.28269958, 31.7918129],
-              [-81.276929529407354, 31.792775687365662],
-              [-81.274276996403628, 31.791644460055245],
-              [-81.271780494753045, 31.788679864345188],
-              [-81.267798025165376, 31.784518630823097],
-              [-81.271030963829091, 31.767815114393926],
-              [-81.274263902492805, 31.762606490991281],
-              [-81.27809906, 31.76241112],
-              [-81.27689362, 31.75951195],
-              [-81.27359772, 31.75701141],
-              [-81.269428590285528, 31.756906212807301],
-              [-81.265316440646416, 31.753753564750639],
-              [-81.26269531, 31.75061226],
-              [-81.25989611, 31.74971224],
-              [-81.2552966, 31.74771216],
-              [-81.2508966, 31.74681216],
-              [-81.2465966, 31.74691216],
-              [-81.231794586947984, 31.743496358761359],
-              [-81.220237036646026, 31.741193705061352],
-              [-81.209210867967144, 31.734861407386333],
-              [-81.1541966, 31.69521216],
-              [-81.15379658, 31.693578],
-              [-81.15389701, 31.69213849],
-              [-81.1537631, 31.69173677],
-              [-81.15399744, 31.69116766],
-              [-81.15383005, 31.69039769],
-              [-81.15393048, 31.68992902],
-              [-81.15359571, 31.68976163],
-              [-81.15409787, 31.68909209],
-              [-81.15389701, 31.68815474],
-              [-81.15416482, 31.68728434],
-              [-81.15456655, 31.68611265],
-              [-81.15496827, 31.68561049],
-              [-81.15530304, 31.68397012],
-              [-81.1559391, 31.68232976],
-              [-81.15617344, 31.68169369],
-              [-81.15650821, 31.68109111],
-              [-81.15654168, 31.68038809],
-              [-81.15684298, 31.67894859],
-              [-81.15704384, 31.67727474],
-              [-81.15747904, 31.67533308],
-              [-81.15784728, 31.67422834],
-              [-81.15821553, 31.67369271],
-              [-81.15851682, 31.6722532],
-              [-81.15905245, 31.67104803],
-              [-81.15955461, 31.66937418],
-              [-81.16049196, 31.66759991],
-              [-81.16106107, 31.66689689],
-              [-81.16173061, 31.66642821],
-              [-81.16209885, 31.66582563],
-              [-81.16383965, 31.66512261],
-              [-81.16460962, 31.66482132],
-              [-81.1663839, 31.66478785],
-              [-81.16701996, 31.66468741],
-              [-81.16835904, 31.66334834],
-              [-81.16775646, 31.6626788],
-              [-81.16789036, 31.66186732],
-              [-81.16842599, 31.66169592],
-              [-81.16861882, 31.66088176],
-              [-81.17084705, 31.65901776],
-              [-81.17279674, 31.65569685],
-              [-81.173868, 31.65501124],
-              [-81.17498747, 31.65340569],
-              [-81.17562353, 31.65270267],
-              [-81.17662784, 31.6504932],
-              [-81.17833516, 31.64848458],
-              [-81.17940643, 31.6474468],
-              [-81.18151547, 31.64563904],
-              [-81.18225196, 31.64500298],
-              [-81.1833567, 31.64296089],
-              [-81.18479621, 31.64125357],
-              [-81.185834, 31.64058403],
-              [-81.18767523, 31.63968015],
-              [-81.18794304, 31.63897713],
-              [-81.18774218, 31.63854193],
-              [-81.18811043, 31.63690156],
-              [-81.18817738, 31.63649984],
-              [-81.1884452, 31.63609812],
-              [-81.18821086, 31.63546206],
-              [-81.18837824, 31.6349599],
-              [-81.18774218, 31.63378821],
-              [-81.18660397, 31.63251609],
-              [-81.18613529, 31.63137787],
-              [-81.18623572, 31.63007227],
-              [-81.18593443, 31.62913492],
-              [-81.18616877, 31.62796323],
-              [-81.18590095, 31.62672458],
-              [-81.18506403, 31.62602156],
-              [-81.18439449, 31.6249503],
-              [-81.18342366, 31.62404642],
-              [-81.18255326, 31.62207129],
-              [-81.18164938, 31.62039744],
-              [-81.18128113, 31.61949356],
-              [-81.18124766, 31.61858969],
-              [-81.18097984, 31.61728409],
-              [-81.18102403, 31.61576022],
-              [-81.18113116, 31.61503176],
-              [-81.18134541, 31.61372482],
-              [-81.18173106, 31.61256786],
-              [-81.18181676, 31.6117537],
-              [-81.18190247, 31.61104666],
-              [-81.18188358, 31.61090557],
-              [-81.177435415821506, 31.586416556810573],
-              [-81.203254, 31.616936],
-              [-81.204704, 31.617352],
-              [-81.208656, 31.617497],
-              [-81.211432, 31.618064],
-              [-81.213659, 31.622357],
-              [-81.215307, 31.626383],
-              [-81.218228, 31.630393],
-              [-81.219943, 31.632042],
-              [-81.222809, 31.633531],
-              [-81.2253, 31.634264],
-              [-81.226131, 31.634768],
-              [-81.22806, 31.636486],
-              [-81.231035, 31.638685],
-              [-81.233313, 31.640655],
-              [-81.235617, 31.64219],
-              [-81.240226, 31.643631],
-              [-81.247941, 31.644246],
-              [-81.256566970607253, 31.647993699430778],
-              [-81.261689, 31.65312],
-              [-81.26576, 31.654863],
-              [-81.27294, 31.655965],
-              [-81.278325, 31.655577],
-              [-81.283281, 31.654593],
-              [-81.290087, 31.653884],
-              [-81.291561, 31.653426],
-              [-81.291802, 31.653082],
-              [-81.2929, 31.652693],
-              [-81.294508, 31.652464],
-              [-81.29701312, 31.65257651],
-              [-81.29849435, 31.64938182],
-              [-81.30241394, 31.65071678],
-              [-81.308087, 31.655183],
-              [-81.30950165, 31.65649223],
-              [-81.3128891, 31.65867615],
-              [-81.31454468, 31.65938187],
-              [-81.31801605, 31.65985489],
-              [-81.31963348, 31.65945244],
-              [-81.32054138, 31.65881538],
-              [-81.325112, 31.654589],
-              [-81.326846, 31.653516],
-              [-81.330393, 31.652109],
-              [-81.335122, 31.650601],
-              [-81.3356781, 31.65026665],
-              [-81.33626598, 31.64825207],
-              [-81.337015, 31.647514],
-              [-81.33808082, 31.6471121],
-              [-81.34016418, 31.6473484],
-              [-81.342215, 31.649732],
-              [-81.344422, 31.651612],
-              [-81.349584, 31.653628],
-              [-81.35557556, 31.65467072],
-              [-81.35679626, 31.65433502],
-              [-81.35896301, 31.65282631],
-              [-81.35990906, 31.65249062],
-              [-81.360578, 31.652625],
-              [-81.36148071, 31.65383339],
-              [-81.362351, 31.654404],
-              [-81.36294556, 31.65453911],
-              [-81.364322, 31.654337],
-              [-81.364755, 31.653532],
-              [-81.36460113, 31.65037727],
-              [-81.364755, 31.649974],
-              [-81.36558533, 31.64920235],
-              [-81.366174, 31.648968],
-              [-81.367789, 31.648968],
-              [-81.36830167, 31.64937072],
-              [-81.36959839, 31.65175438],
-              [-81.37003326, 31.65195656],
-              [-81.37094116, 31.65178871],
-              [-81.37153625, 31.65138626],
-              [-81.371651, 31.650614],
-              [-81.371178, 31.64907],
-              [-81.371209, 31.648263],
-              [-81.371365, 31.64806],
-              [-81.37239075, 31.64761734],
-              [-81.37440491, 31.64717102],
-              [-81.37482452, 31.64730072],
-              [-81.37552643, 31.64785385],
-              [-81.37562561, 31.64831352],
-              [-81.3755188, 31.64854622],
-              [-81.37482452, 31.64906502],
-              [-81.373542, 31.649204],
-              [-81.37310791, 31.64940643],
-              [-81.372794, 31.649842],
-              [-81.37271118, 31.65034485],
-              [-81.372991, 31.650782],
-              [-81.374173, 31.651386],
-              [-81.37456512, 31.65175438],
-              [-81.374607, 31.653769],
-              [-81.37615, 31.654707],
-              [-81.377776, 31.655106],
-              [-81.378061, 31.654947],
-              [-81.378132, 31.654496],
-              [-81.377901, 31.654083],
-              [-81.377216, 31.653828],
-              [-81.376552, 31.653308],
-              [-81.376297, 31.6527977],
-              [-81.37628174, 31.6524601],
-              [-81.37667847, 31.65155983],
-              [-81.37676298, 31.65050609],
-              [-81.37698364, 31.650177],
-              [-81.37782267, 31.65016743],
-              [-81.37923431, 31.65092659],
-              [-81.37985992, 31.65101242],
-              [-81.380265, 31.650862],
-              [-81.38043213, 31.65045357],
-              [-81.380161, 31.64925],
-              [-81.38021851, 31.64865494],
-              [-81.38053131, 31.64840698],
-              [-81.38098144, 31.64837837],
-              [-81.38238, 31.64888],
-              [-81.38293447, 31.64736957],
-              [-81.38378906, 31.64681244],
-              [-81.384976, 31.646845],
-              [-81.38669586, 31.64884758],
-              [-81.38720703, 31.64882851],
-              [-81.3874054, 31.64869118],
-              [-81.38709259, 31.64818955],
-              [-81.38709246, 31.64781507],
-              [-81.3877182, 31.64694595],
-              [-81.38799286, 31.64622307],
-              [-81.38857269, 31.64613915],
-              [-81.38905334, 31.6464901],
-              [-81.389328, 31.64648247],
-              [-81.38945007, 31.64561462],
-              [-81.389817, 31.645415],
-              [-81.39019775, 31.64570808],
-              [-81.390556, 31.646327],
-              [-81.39079285, 31.64639282],
-              [-81.39174652, 31.64560318],
-              [-81.39186096, 31.64533043],
-              [-81.390803, 31.644847],
-              [-81.390633, 31.644558],
-              [-81.390756, 31.644228],
-              [-81.391482, 31.64389],
-              [-81.39255524, 31.64370537],
-              [-81.393672, 31.643798],
-              [-81.3931186, 31.64359621],
-              [-81.3963966, 31.64351216],
-              [-81.4010966, 31.64571216],
-              [-81.4040966, 31.64731216],
-              [-81.4072966, 31.64841216],
-              [-81.4099966, 31.64751216],
-              [-81.4116966, 31.64881216],
-              [-81.4142966, 31.65001216],
-              [-81.4162966, 31.65211216],
-              [-81.4212966, 31.65041216],
-              [-81.4246966, 31.64901216],
-              [-81.426642, 31.64529],
-              [-81.426627, 31.644717],
-              [-81.426927, 31.643717],
-              [-81.427827, 31.642617],
-              [-81.428127, 31.642517],
-              [-81.429527, 31.641018],
-              [-81.431127, 31.640518],
-              [-81.432127, 31.640518],
-              [-81.434527, 31.641317],
-              [-81.434927, 31.641717],
-              [-81.436928, 31.642917],
-              [-81.437728, 31.643617],
-              [-81.438528, 31.645517],
-              [-81.439528, 31.647117],
-              [-81.441828, 31.653417],
-              [-81.442028, 31.653817],
-              [-81.443128, 31.654917],
-              [-81.445728, 31.656217],
-              [-81.448728, 31.656817],
-              [-81.451928, 31.657817],
-              [-81.452928, 31.658417],
-              [-81.456629, 31.661317],
-              [-81.456629, 31.661517],
-              [-81.457529, 31.662817],
-              [-81.461529, 31.672416],
-              [-81.462329, 31.673316],
-              [-81.465629, 31.676316],
-              [-81.467929, 31.678916],
-              [-81.468029, 31.679616],
-              [-81.467129, 31.680616],
-              [-81.466529, 31.680916],
-              [-81.464529, 31.681316],
-              [-81.461839, 31.682349],
-              [-81.461101, 31.683215],
-              [-81.46104, 31.684044],
-              [-81.461263, 31.685211],
-              [-81.461129, 31.687216],
-              [-81.462329, 31.688916],
-              [-81.463529, 31.690016],
-              [-81.464229, 31.690416],
-              [-81.464629, 31.690916],
-              [-81.464929, 31.692416],
-              [-81.465629, 31.693616],
-              [-81.46873, 31.695216],
-              [-81.47063, 31.695816],
-              [-81.47453, 31.697816],
-              [-81.47663, 31.698516],
-              [-81.48243, 31.699316],
-              [-81.48493, 31.699315],
-              [-81.48633, 31.699515],
-              [-81.49186805, 31.69941952],
-              [-81.49979401, 31.70001221],
-              [-81.5959966, 31.74451216],
-              [-81.6263966, 31.75911216],
-              [-81.6417966, 31.76571216],
-              [-81.64339447, 31.78541184],
-              [-81.64609528, 31.78831291],
-              [-81.64789581, 31.78971291],
-              [-81.65999603, 31.80051231],
-              [-81.6625966, 31.80151216],
-              [-81.6661966, 31.80261216],
-              [-81.6864966, 31.80871216],
-              [-81.68889618, 31.80931282],
-              [-81.69539642, 31.8135128],
-              [-81.6969966, 31.81721216],
-              [-81.69909668, 31.82001305],
-              [-81.7052966, 31.82641216],
-              [-81.70689392, 31.82831192],
-              [-81.7147966, 31.83541216],
-              [-81.71779633, 31.8366127],
-              [-81.7284966, 31.84191216],
-              [-81.73569488, 31.85391235],
-              [-81.7482966, 31.86071216],
-              [-81.75329605, 31.86261195],
-              [-81.7602966, 31.87051216],
-              [-81.7650966, 31.87461216],
-              [-81.7658966, 31.88091216],
-              [-81.7659966, 31.88641216],
-              [-81.76439667, 31.89061165],
-              [-81.76259613, 31.89421272],
-              [-81.7586966, 31.89961216],
-              [-81.75889587, 31.90391159],
-              [-81.7595966, 31.90611216],
-              [-81.7528966, 31.91491216],
-              [-81.7508966, 31.92091216],
-              [-81.74949646, 31.92621231],
-              [-81.7483966, 31.92941216],
-              [-81.74609375, 31.93411255],
-              [-81.7453966, 31.93771216],
-              [-81.7461966, 31.94521216],
-              [-81.7453966, 31.94831216],
-              [-81.74329376, 31.95341301],
-              [-81.7428966, 31.95621216],
-              [-81.7435966, 31.96091216],
-              [-81.7464966, 31.96821216],
-              [-81.74969482, 31.97281265],
-              [-81.7536966, 31.97461216],
-              [-81.77129374, 31.97761165],
-              [-81.7736966, 31.98021216],
-              [-81.77759552, 31.98101234],
-              [-81.78139496, 31.98051262],
-              [-81.78779602, 31.97881317],
-              [-81.79409684, 31.97841269],
-              [-81.79529571, 31.98111153],
-              [-81.79829407, 31.98141289],
-              [-81.8014966, 31.98401216],
-              [-81.8041966, 31.98701216],
-              [-81.80489349, 31.99341202],
-              [-81.8079966, 31.99741216],
-              [-81.8099966, 31.99871216],
-              [-81.8125966, 32.00161216],
-              [-81.8184967, 32.00441361],
-              [-81.82249451, 32.00741196],
-              [-81.82439423, 32.01371384],
-              [-81.82069397, 32.01521301],
-              [-81.81399536, 32.01611328],
-              [-81.8100966, 32.01961216],
-              [-81.76169614, 32.0479126],
-              [-81.74079895, 32.06781387],
-              [-81.72019958, 32.08771133],
-              [-81.71769836, 32.08931273],
-              [-81.7149966, 32.09031216],
-              [-81.7141966, 32.09221216],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        id: "InlandCounties.2",
-        STATE: "GA",
-        CWA: "JAX",
-        COUNTYNAME: "Camden",
-        FIPS: "13039",
-        TIME_ZONE: "E",
-        FE_AREA: "se",
-        LON: -81.682,
-        LAT: 30.9335,
-        uuid: "66bc87d3-4ae2-5b56-b2b4-f9146acce491",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-81.56349945, 31.08862114],
-              [-81.56371174, 31.08848467],
-              [-81.56445312, 31.08866692],
-              [-81.56535997, 31.09063791],
-              [-81.565254, 31.090707],
-              [-81.564668, 31.090318],
-              [-81.56400299, 31.08969879],
-              [-81.56349945, 31.08862114],
-            ],
-          ],
-          [
-            [
-              [-81.76269531, 31.15371323],
-              [-81.75429535, 31.1505127],
-              [-81.7508316, 31.1505127],
-              [-81.74723053, 31.15101242],
-              [-81.74443054, 31.1511116],
-              [-81.74032593, 31.15151215],
-              [-81.73580933, 31.15161133],
-              [-81.73199463, 31.15071297],
-              [-81.72879791, 31.14961243],
-              [-81.72409821, 31.14951134],
-              [-81.72189331, 31.14861298],
-              [-81.71649933, 31.14451218],
-              [-81.7151947, 31.14171219],
-              [-81.71259308, 31.13851166],
-              [-81.71029663, 31.13671303],
-              [-81.70329285, 31.13421249],
-              [-81.69389343, 31.1274128],
-              [-81.69109344, 31.12571144],
-              [-81.68610382, 31.12231636],
-              [-81.66381836, 31.12521172],
-              [-81.66189575, 31.12511253],
-              [-81.6588974, 31.12401199],
-              [-81.6579361, 31.12235832],
-              [-81.6574173, 31.12250328],
-              [-81.65756226, 31.12205124],
-              [-81.65496826, 31.1213665],
-              [-81.65331268, 31.12129784],
-              [-81.65219116, 31.12104416],
-              [-81.65163422, 31.12072372],
-              [-81.65070343, 31.11966896],
-              [-81.64990234, 31.11820221],
-              [-81.64899445, 31.11701012],
-              [-81.64723969, 31.11558723],
-              [-81.64654541, 31.11558723],
-              [-81.6461792, 31.11581802],
-              [-81.64373016, 31.11799431],
-              [-81.64266968, 31.11831284],
-              [-81.64212036, 31.11815453],
-              [-81.64159393, 31.11758041],
-              [-81.64073944, 31.11574745],
-              [-81.63549805, 31.11125374],
-              [-81.63471985, 31.11038208],
-              [-81.63362885, 31.10873222],
-              [-81.63301849, 31.10829544],
-              [-81.63176727, 31.10862923],
-              [-81.63008881, 31.1099453],
-              [-81.62828064, 31.11177826],
-              [-81.62625122, 31.11475563],
-              [-81.62545013, 31.11541939],
-              [-81.62380981, 31.11649513],
-              [-81.62154388, 31.11739159],
-              [-81.61871338, 31.11769104],
-              [-81.61693573, 31.11750984],
-              [-81.61560821, 31.11663818],
-              [-81.615448, 31.11615753],
-              [-81.61470032, 31.1152401],
-              [-81.61395264, 31.11469078],
-              [-81.61048889, 31.11334038],
-              [-81.60955048, 31.11315918],
-              [-81.60800934, 31.11336327],
-              [-81.6014328, 31.11584282],
-              [-81.59928402, 31.11634636],
-              [-81.59947463, 31.11543393],
-              [-81.598472, 31.115453],
-              [-81.59540558, 31.11494827],
-              [-81.593944, 31.115064],
-              [-81.592932, 31.115454],
-              [-81.589682, 31.117655],
-              [-81.588324, 31.117885],
-              [-81.587605, 31.11761],
-              [-81.586566, 31.116625],
-              [-81.58563232, 31.11602783],
-              [-81.583262, 31.114723],
-              [-81.582197, 31.114517],
-              [-81.58052063, 31.11490631],
-              [-81.578308, 31.115778],
-              [-81.57652283, 31.1172905],
-              [-81.5752182, 31.12026977],
-              [-81.57468414, 31.12382317],
-              [-81.57395172, 31.12523079],
-              [-81.572767, 31.126576],
-              [-81.570237, 31.128547],
-              [-81.56770325, 31.12941742],
-              [-81.5671463, 31.12941742],
-              [-81.564934, 31.128524],
-              [-81.56248474, 31.12806511],
-              [-81.56144714, 31.12776756],
-              [-81.560619, 31.127332],
-              [-81.559899, 31.126691],
-              [-81.559167, 31.12523],
-              [-81.5589447, 31.12352562],
-              [-81.55915832, 31.12162399],
-              [-81.55944824, 31.12066078],
-              [-81.561259, 31.117315],
-              [-81.56182098, 31.11575699],
-              [-81.56264496, 31.11204338],
-              [-81.562671, 31.110989],
-              [-81.56394959, 31.10746002],
-              [-81.564136, 31.107368],
-              [-81.564242, 31.106956],
-              [-81.564881, 31.106016],
-              [-81.56485748, 31.10571861],
-              [-81.56610871, 31.10443307],
-              [-81.566613, 31.104159],
-              [-81.56757355, 31.10395241],
-              [-81.56829071, 31.10349464],
-              [-81.56930459, 31.102395],
-              [-81.569569, 31.102395],
-              [-81.569782, 31.102601],
-              [-81.570421, 31.102532],
-              [-81.57191467, 31.10111046],
-              [-81.57300568, 31.0996666],
-              [-81.57415008, 31.09700775],
-              [-81.574442, 31.096206],
-              [-81.574228, 31.094831],
-              [-81.57366943, 31.09395981],
-              [-81.57202148, 31.09281349],
-              [-81.57095337, 31.09251595],
-              [-81.56661224, 31.09192085],
-              [-81.566026, 31.091646],
-              [-81.565866, 31.090295],
-              [-81.56536102, 31.08875847],
-              [-81.56394959, 31.08701706],
-              [-81.56299, 31.085207],
-              [-81.55830383, 31.08039475],
-              [-81.557078, 31.079752],
-              [-81.555933, 31.079431],
-              [-81.553936, 31.079202],
-              [-81.55268097, 31.07922554],
-              [-81.550953, 31.079775],
-              [-81.548637, 31.081562],
-              [-81.54756928, 31.08213425],
-              [-81.54599943, 31.08250202],
-              [-81.54402924, 31.08256912],
-              [-81.540329, 31.081905],
-              [-81.538119, 31.080942],
-              [-81.536149, 31.079773],
-              [-81.53503418, 31.07876396],
-              [-81.53460693, 31.07805252],
-              [-81.534099, 31.076884],
-              [-81.53394318, 31.0759449],
-              [-81.53375244, 31.0759449],
-              [-81.533619, 31.075738],
-              [-81.533327, 31.07489],
-              [-81.533167, 31.074294],
-              [-81.5337, 31.073584],
-              [-81.53372955, 31.07253075],
-              [-81.533488, 31.071017],
-              [-81.53237152, 31.06803703],
-              [-81.53260894, 31.06790061],
-              [-81.53077698, 31.06485176],
-              [-81.53079991, 31.06466863],
-              [-81.52877808, 31.06331444],
-              [-81.52680969, 31.0620079],
-              [-81.524066, 31.060678],
-              [-81.51821136, 31.05840874],
-              [-81.498376937174072, 31.055487829951126],
-              [-81.491414156059307, 31.051832369865881],
-              [-81.490485785244019, 31.046494237677901],
-              [-81.486308116575159, 31.033845185319425],
-              [-81.48427582, 31.0281086],
-              [-81.450671554384328, 31.030343233204697],
-              [-81.449467336777133, 30.972735060134305],
-              [-81.442136312772007, 30.949232659647254],
-              [-81.48457336, 30.84535599],
-              [-81.48457336, 30.844347],
-              [-81.478709948166994, 30.769584418255384],
-              [-81.481151497809975, 30.75295136131253],
-              [-81.49656377993135, 30.730061833409515],
-              [-81.5026778, 30.72773372],
-              [-81.510665, 30.724368],
-              [-81.511832, 30.724162],
-              [-81.515015, 30.724392],
-              [-81.517801, 30.724852],
-              [-81.52002925, 30.72489799],
-              [-81.521542, 30.725334],
-              [-81.522524, 30.725426],
-              [-81.52586365, 30.72538185],
-              [-81.53098297, 30.72657585],
-              [-81.535384292126835, 30.717851073103361],
-              [-81.535414702406129, 30.713715275118357],
-              [-81.536296, 30.711448],
-              [-81.536508, 30.71115],
-              [-81.536959, 30.710944],
-              [-81.53892363, 30.71087652],
-              [-81.541601, 30.711908],
-              [-81.542609, 30.712802],
-              [-81.543167, 30.713673],
-              [-81.54324341, 30.71509361],
-              [-81.54358673, 30.71598816],
-              [-81.54486084, 30.71802902],
-              [-81.54698181, 30.71928978],
-              [-81.54947662, 30.71974754],
-              [-81.55203344, 30.71949251],
-              [-81.554546, 30.71851],
-              [-81.55555725, 30.7176857],
-              [-81.556535, 30.716585],
-              [-81.55743408, 30.71512413],
-              [-81.557968, 30.714545],
-              [-81.5596695, 30.71326065],
-              [-81.56104444, 30.71257428],
-              [-81.56186676, 30.71271133],
-              [-81.5623169, 30.71319199],
-              [-81.565789060300759, 30.718682129212553],
-              [-81.573918483337138, 30.72263070611594],
-              [-81.585444843856592, 30.723704951302892],
-              [-81.59349928842316, 30.718956304287452],
-              [-81.596999171271236, 30.722822171590597],
-              [-81.597632272503972, 30.728953257212922],
-              [-81.60462970718163, 30.729153183917997],
-              [-81.606095836352182, 30.719123527546692],
-              [-81.607528644405221, 30.716757728203298],
-              [-81.609249577442711, 30.716005782562103],
-              [-81.611379522649969, 30.716734448027744],
-              [-81.61328125, 30.72207451],
-              [-81.619040801970144, 30.722695932799894],
-              [-81.623089803696644, 30.723643571501842],
-              [-81.623154415426313, 30.726098817229616],
-              [-81.620684342251195, 30.730614043231181],
-              [-81.623155101479298, 30.735792483805142],
-              [-81.627216623498086, 30.735860175838788],
-              [-81.629112000440188, 30.733660184745276],
-              [-81.628739694255131, 30.728887896373191],
-              [-81.632191987971112, 30.728718666289076],
-              [-81.636094742605408, 30.733683939288913],
-              [-81.63988094336186, 30.73131756381613],
-              [-81.643485115235791, 30.727895420824726],
-              [-81.649455662582497, 30.727531363059683],
-              [-81.653155412605258, 30.73010617212601],
-              [-81.653990218778631, 30.733445396819523],
-              [-81.653318468349312, 30.737498666422876],
-              [-81.652669970013775, 30.743844685849197],
-              [-81.654754428949431, 30.744400541565373],
-              [-81.660035058253087, 30.742871938345893],
-              [-81.662675372904914, 30.746207072642939],
-              [-81.659538539323137, 30.750273389734851],
-              [-81.661089798015965, 30.753801742840103],
-              [-81.664405233261419, 30.754501330093731],
-              [-81.670123598638895, 30.751520480056534],
-              [-81.67818406047418, 30.751277145359623],
-              [-81.682746586041318, 30.748935048901824],
-              [-81.683263672272261, 30.746288784072885],
-              [-81.680061918730971, 30.744551797310351],
-              [-81.676274219648164, 30.746390486185497],
-              [-81.674251661885506, 30.748192401283141],
-              [-81.672376199232843, 30.748486591503163],
-              [-81.670942021910236, 30.747714342175602],
-              [-81.668662047705055, 30.744662118642857],
-              [-81.669066559257587, 30.741131836002577],
-              [-81.672817484562884, 30.738741540464886],
-              [-81.678701288963353, 30.741389252445096],
-              [-81.68763039, 30.74080636],
-              [-81.68965912, 30.74103737],
-              [-81.69064331, 30.74138069],
-              [-81.69112396, 30.74172592],
-              [-81.6916275, 30.7425499],
-              [-81.69223785, 30.74316978],
-              [-81.69281769, 30.74568939],
-              [-81.69340515, 30.74690437],
-              [-81.69382477, 30.74743271],
-              [-81.6942749, 30.74761581],
-              [-81.69515228, 30.74756813],
-              [-81.69637299, 30.7470417],
-              [-81.69809723, 30.74600983],
-              [-81.70043182, 30.74502563],
-              [-81.70220947, 30.74465752],
-              [-81.70425415, 30.7446804],
-              [-81.70640564, 30.74562073],
-              [-81.70809936, 30.74669647],
-              [-81.71089172, 30.74676514],
-              [-81.71131134, 30.74603081],
-              [-81.71123505, 30.74415207],
-              [-81.70982361, 30.74218178],
-              [-81.70983124, 30.74160767],
-              [-81.71046448, 30.74144745],
-              [-81.71043396, 30.74225044],
-              [-81.71165466, 30.74394417],
-              [-81.71192169, 30.74458695],
-              [-81.7119751, 30.74653435],
-              [-81.71221924, 30.7466507],
-              [-81.71542358, 30.74570847],
-              [-81.71728516, 30.74458504],
-              [-81.71949005, 30.74401093],
-              [-81.72078705, 30.74403572],
-              [-81.72238159, 30.74483681],
-              [-81.72319794, 30.74556923],
-              [-81.72381592, 30.74586868],
-              [-81.72640991, 30.74593544],
-              [-81.72811127, 30.74648476],
-              [-81.73012543, 30.74779129],
-              [-81.73267364, 30.74914169],
-              [-81.73380279, 30.7502346],
-              [-81.73421478, 30.75094986],
-              [-81.73421478, 30.7520504],
-              [-81.73275757, 30.7532196],
-              [-81.73252106, 30.75404549],
-              [-81.73270416, 30.75436592],
-              [-81.73326111, 30.7548008],
-              [-81.73652649, 30.75567055],
-              [-81.73989868, 30.7569294],
-              [-81.74239349, 30.75816727],
-              [-81.74316406, 30.75885391],
-              [-81.743721, 30.75967979],
-              [-81.743721, 30.76089287],
-              [-81.74308777, 30.7620163],
-              [-81.74242401, 30.76245117],
-              [-81.74141693, 30.7635994],
-              [-81.74151611, 30.76401138],
-              [-81.74186706, 30.76442337],
-              [-81.7436676, 30.76534081],
-              [-81.74613953, 30.76602554],
-              [-81.75104523, 30.76627731],
-              [-81.75370026, 30.7673111],
-              [-81.75723267, 30.76907539],
-              [-81.75831604, 30.76969528],
-              [-81.75964355, 30.77098084],
-              [-81.76091766, 30.77308845],
-              [-81.76147461, 30.77450943],
-              [-81.76301575, 30.77503777],
-              [-81.76691437, 30.77565765],
-              [-81.76890564, 30.77620888],
-              [-81.7694397, 30.77625465],
-              [-81.77020264, 30.77611732],
-              [-81.77036285, 30.77595711],
-              [-81.77039337, 30.77547646],
-              [-81.7696228, 30.77419281],
-              [-81.7696228, 30.77343559],
-              [-81.76996613, 30.77265739],
-              [-81.77159119, 30.77151108],
-              [-81.77193451, 30.76974678],
-              [-81.7730484, 30.76779938],
-              [-81.77302551, 30.76704216],
-              [-81.77106476, 30.76438332],
-              [-81.77114105, 30.76369667],
-              [-81.77148437, 30.76335335],
-              [-81.77243805, 30.76298523],
-              [-81.7800293, 30.76108551],
-              [-81.78189087, 30.76081276],
-              [-81.78281403, 30.76090431],
-              [-81.78385162, 30.76133919],
-              [-81.78433227, 30.76186562],
-              [-81.78424835, 30.76312637],
-              [-81.78388214, 30.76434135],
-              [-81.7835083, 30.76498222],
-              [-81.78353119, 30.76569367],
-              [-81.78449249, 30.76683998],
-              [-81.78557587, 30.76775551],
-              [-81.78674316, 30.76842117],
-              [-81.78751373, 30.76931572],
-              [-81.78740692, 30.76988792],
-              [-81.78679657, 30.77066803],
-              [-81.78618622, 30.77119446],
-              [-81.78507233, 30.77165222],
-              [-81.78456879, 30.77213287],
-              [-81.78459168, 30.7728672],
-              [-81.78679657, 30.77596092],
-              [-81.78844452, 30.777565],
-              [-81.79263306, 30.78320503],
-              [-81.79329681, 30.78474045],
-              [-81.79340362, 30.78609276],
-              [-81.79361725, 30.78650475],
-              [-81.79430389, 30.78636742],
-              [-81.79595184, 30.78554344],
-              [-81.79693604, 30.78478622],
-              [-81.79711914, 30.78416824],
-              [-81.79677582, 30.78341103],
-              [-81.79584503, 30.78233337],
-              [-81.79587555, 30.7815094],
-              [-81.79611206, 30.78120995],
-              [-81.79677582, 30.78088951],
-              [-81.79804993, 30.78086853],
-              [-81.79918671, 30.78182983],
-              [-81.80097198, 30.78375626],
-              [-81.80162811, 30.78469467],
-              [-81.80218506, 30.78595543],
-              [-81.80269623, 30.78648376],
-              [-81.8049469, 30.78774452],
-              [-81.80630493, 30.78907394],
-              [-81.80699158, 30.7893486],
-              [-81.8092804, 30.78962326],
-              [-81.81137085, 30.78948593],
-              [-81.81453705, 30.78882027],
-              [-81.81944275, 30.78918839],
-              [-81.82039642, 30.78898048],
-              [-81.82236481, 30.78767586],
-              [-81.82390595, 30.786232],
-              [-81.82417297, 30.78616142],
-              [-81.82563019, 30.78671265],
-              [-81.82589722, 30.78755951],
-              [-81.827034, 30.78838539],
-              [-81.82977295, 30.78868294],
-              [-81.83090973, 30.78858948],
-              [-81.83361816, 30.78799438],
-              [-81.83648682, 30.78687096],
-              [-81.83916473, 30.78611374],
-              [-81.84046936, 30.78602219],
-              [-81.8421402, 30.78654862],
-              [-81.84264374, 30.78721237],
-              [-81.84264374, 30.78792381],
-              [-81.84248352, 30.78831291],
-              [-81.84101868, 30.79021835],
-              [-81.84092712, 30.79091263],
-              [-81.84247589, 30.79103661],
-              [-81.84479523, 30.79014587],
-              [-81.84673309, 30.79035187],
-              [-81.84840393, 30.79103851],
-              [-81.85063934, 30.79239082],
-              [-81.85250854, 30.79387665],
-              [-81.85307312, 30.79404259],
-              [-81.85366821, 30.79390144],
-              [-81.85453796, 30.79337502],
-              [-81.85700989, 30.79362488],
-              [-81.85826111, 30.79335022],
-              [-81.85958862, 30.79332733],
-              [-81.86314392, 30.79362488],
-              [-81.86765289, 30.79252243],
-              [-81.8688736, 30.79236031],
-              [-81.86993408, 30.7929554],
-              [-81.87052155, 30.79364204],
-              [-81.8711319, 30.79476547],
-              [-81.87121582, 30.79499435],
-              [-81.87091827, 30.7964859],
-              [-81.87110901, 30.79694367],
-              [-81.87145233, 30.79721832],
-              [-81.87360382, 30.79781342],
-              [-81.87593079, 30.79884529],
-              [-81.88298798, 30.79930687],
-              [-81.88365173, 30.79972076],
-              [-81.88378143, 30.80056763],
-              [-81.88359833, 30.80205727],
-              [-81.88304138, 30.80407524],
-              [-81.88314819, 30.80476189],
-              [-81.8833847, 30.80508232],
-              [-81.88497925, 30.80613708],
-              [-81.88537598, 30.80700874],
-              [-81.88545227, 30.8108139],
-              [-81.88577271, 30.81108856],
-              [-81.88807678, 30.81177711],
-              [-81.88935089, 30.81262588],
-              [-81.89051819, 30.81363487],
-              [-81.89094543, 30.81461906],
-              [-81.89219666, 30.81622505],
-              [-81.89211273, 30.81760025],
-              [-81.89173889, 30.81830978],
-              [-81.89216614, 30.82050896],
-              [-81.89182282, 30.82135773],
-              [-81.8918457, 30.82261848],
-              [-81.89342499, 30.82476807],
-              [-81.89561462, 30.82452011],
-              [-81.89672852, 30.82493401],
-              [-81.90039825, 30.82800674],
-              [-81.90084076, 30.82857895],
-              [-81.90161133, 30.8290844],
-              [-81.90267181, 30.82896233],
-              [-81.90477753, 30.82777786],
-              [-81.90509796, 30.82800865],
-              [-81.90499878, 30.82852936],
-              [-81.90461731, 30.82870674],
-              [-81.90422821, 30.83351326],
-              [-81.90312958, 30.84081268],
-              [-81.89662933, 30.88161278],
-              [-81.89592743, 30.88831139],
-              [-81.89402771, 30.90001297],
-              [-81.89411926, 30.9056282],
-              [-81.89424133, 30.90514946],
-              [-81.89553833, 30.90484428],
-              [-81.89680481, 30.90494156],
-              [-81.89742279, 30.9052105],
-              [-81.89800263, 30.90684318],
-              [-81.89990997, 30.90859413],
-              [-81.90153503, 30.90850449],
-              [-81.90480804, 30.90796089],
-              [-81.90561676, 30.90795708],
-              [-81.9061203, 30.90822792],
-              [-81.90643311, 30.90863037],
-              [-81.90644073, 30.90920258],
-              [-81.90616608, 30.90970993],
-              [-81.90410614, 30.911129],
-              [-81.90367127, 30.91168594],
-              [-81.90260315, 30.91241264],
-              [-81.90142822, 30.91406441],
-              [-81.90154266, 30.91435432],
-              [-81.90285492, 30.91520309],
-              [-81.90393829, 30.91524124],
-              [-81.90697479, 30.91443443],
-              [-81.90813446, 30.91332626],
-              [-81.90925598, 30.91169167],
-              [-81.91003418, 30.91084862],
-              [-81.91170502, 30.90979958],
-              [-81.91214752, 30.90969849],
-              [-81.91435242, 30.91010666],
-              [-81.9154892, 30.91127586],
-              [-81.91567993, 30.91161728],
-              [-81.91595459, 30.9131279],
-              [-81.91588592, 30.91535377],
-              [-81.91477203, 30.91746139],
-              [-81.91287994, 30.92172432],
-              [-81.90992737, 30.92738533],
-              [-81.90799713, 30.9286499],
-              [-81.90555572, 30.9317894],
-              [-81.90408325, 30.93438721],
-              [-81.90246582, 30.93796349],
-              [-81.90122986, 30.94019127],
-              [-81.90129089, 30.94072151],
-              [-81.90010834, 30.94448662],
-              [-81.89813995, 30.94710541],
-              [-81.89626312, 30.94795036],
-              [-81.89448547, 30.94952202],
-              [-81.8941803, 30.94957924],
-              [-81.89315796, 30.9507904],
-              [-81.89274597, 30.95254326],
-              [-81.89272308, 30.95476723],
-              [-81.89225769, 30.96055794],
-              [-81.89420319, 30.96160126],
-              [-81.89756012, 30.96216965],
-              [-81.89974976, 30.96294022],
-              [-81.90201569, 30.96519852],
-              [-81.90311432, 30.9657383],
-              [-81.90377808, 30.96657944],
-              [-81.90381622, 30.96712112],
-              [-81.90366363, 30.96745873],
-              [-81.90255737, 30.96821213],
-              [-81.90222931, 30.96931267],
-              [-81.90487671, 30.97029304],
-              [-81.90542603, 30.97056198],
-              [-81.90632629, 30.97150612],
-              [-81.90718079, 30.97167396],
-              [-81.90776825, 30.97198296],
-              [-81.90744781, 30.97287559],
-              [-81.90689087, 30.97352219],
-              [-81.90722656, 30.97489357],
-              [-81.90991211, 30.97311401],
-              [-81.91202545, 30.97331238],
-              [-81.91413117, 30.9752121],
-              [-81.91672516, 30.98191261],
-              [-81.91752625, 30.98491287],
-              [-81.91742706, 30.98781204],
-              [-81.91492462, 30.98981285],
-              [-81.91272736, 30.99191284],
-              [-81.91392517, 30.99530601],
-              [-81.91110992, 30.99481583],
-              [-81.90792847, 30.99541283],
-              [-81.90722656, 30.99931145],
-              [-81.91072846, 31.00381279],
-              [-81.90992737, 31.00741196],
-              [-81.91053009, 31.01101303],
-              [-81.90953064, 31.01481247],
-              [-81.90742493, 31.01691246],
-              [-81.90592956, 31.0200119],
-              [-81.90473175, 31.02400971],
-              [-81.90752411, 31.02371216],
-              [-81.90792847, 31.02591324],
-              [-81.90592956, 31.02781296],
-              [-81.90342712, 31.02871132],
-              [-81.90532684, 31.02981186],
-              [-81.90532684, 31.03311157],
-              [-81.90602875, 31.03511238],
-              [-81.90532684, 31.03721237],
-              [-81.90712738, 31.04001236],
-              [-81.91082764, 31.04211235],
-              [-81.91513061, 31.04261208],
-              [-81.91811371, 31.04370689],
-              [-81.92131043, 31.04241371],
-              [-81.92422485, 31.04321098],
-              [-81.92592621, 31.04521179],
-              [-81.92832947, 31.04681206],
-              [-81.93152618, 31.04711151],
-              [-81.93512726, 31.04851532],
-              [-81.93602753, 31.05231285],
-              [-81.93332672, 31.05431175],
-              [-81.932724, 31.05691147],
-              [-81.93612671, 31.05981255],
-              [-81.93402862, 31.0628128],
-              [-81.93131256, 31.06331062],
-              [-81.92959595, 31.06131172],
-              [-81.92712402, 31.06111527],
-              [-81.92692566, 31.06521225],
-              [-81.92633057, 31.06941223],
-              [-81.92682648, 31.07291222],
-              [-81.92682648, 31.07951164],
-              [-81.92472839, 31.08171272],
-              [-81.92453003, 31.08551216],
-              [-81.92272949, 31.08891106],
-              [-81.92022705, 31.09091187],
-              [-81.91579437, 31.09521294],
-              [-81.91332245, 31.09471703],
-              [-81.91132355, 31.09661293],
-              [-81.90782929, 31.09761238],
-              [-81.89822388, 31.09961128],
-              [-81.8881073, 31.10041046],
-              [-81.88579559, 31.09941101],
-              [-81.88150024, 31.0988121],
-              [-81.84880829, 31.09791184],
-              [-81.84092712, 31.09801292],
-              [-81.83792877, 31.10311317],
-              [-81.83272553, 31.10851288],
-              [-81.83242798, 31.11471176],
-              [-81.83132935, 31.1172123],
-              [-81.82902527, 31.12031174],
-              [-81.8248291, 31.12461281],
-              [-81.8228302, 31.12971115],
-              [-81.81852722, 31.13671303],
-              [-81.8133316, 31.14051247],
-              [-81.80221558, 31.15051079],
-              [-81.79012299, 31.14682007],
-              [-81.78872681, 31.14931297],
-              [-81.78812408, 31.15171242],
-              [-81.78142548, 31.15761185],
-              [-81.77492523, 31.16281128],
-              [-81.77302551, 31.16531181],
-              [-81.76832581, 31.16901207],
-              [-81.76519775, 31.16921234],
-              [-81.76309967, 31.16671181],
-              [-81.76309967, 31.16341209],
-              [-81.76439667, 31.16001129],
-              [-81.76439667, 31.15651131],
-              [-81.76269531, 31.15371323],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        id: "InlandCounties.3",
-        STATE: "GA",
-        CWA: "CHS",
-        COUNTYNAME: "Bryan",
-        FIPS: "13029",
-        TIME_ZONE: "E",
-        FE_AREA: "se",
-        LON: -81.4474,
-        LAT: 32.0175,
-        uuid: "8fa8c48f-4eb6-5471-9d9f-dad09d0e7cb3",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-81.4221966, 32.21281216],
-              [-81.4201966, 32.21081216],
-              [-81.4159966, 32.21131216],
-              [-81.4142966, 32.20841216],
-              [-81.4134966, 32.20551216],
-              [-81.4156966, 32.20291216],
-              [-81.4160966, 32.19921216],
-              [-81.4183966, 32.19681216],
-              [-81.4163966, 32.19311216],
-              [-81.4150966, 32.18841216],
-              [-81.4117966, 32.18571216],
-              [-81.4140966, 32.18381216],
-              [-81.4146966, 32.18021216],
-              [-81.4132966, 32.17611216],
-              [-81.4136966, 32.17321216],
-              [-81.4127966, 32.17021216],
-              [-81.4104966, 32.16871216],
-              [-81.4113966, 32.16461216],
-              [-81.4110966, 32.16141216],
-              [-81.4132966, 32.16071216],
-              [-81.4107966, 32.15731216],
-              [-81.4079966, 32.15441216],
-              [-81.4067966, 32.15191216],
-              [-81.4025966, 32.15031216],
-              [-81.4029966, 32.14761216],
-              [-81.4024966, 32.14431216],
-              [-81.4040966, 32.14321216],
-              [-81.4055966, 32.14101216],
-              [-81.4067966, 32.13771216],
-              [-81.4068966, 32.13561216],
-              [-81.4062966, 32.13201216],
-              [-81.4064966, 32.12911216],
-              [-81.4079966, 32.12731216],
-              [-81.4062966, 32.12581216],
-              [-81.4051966, 32.12321216],
-              [-81.4061966, 32.12051216],
-              [-81.4048966, 32.11531216],
-              [-81.4032966, 32.11361216],
-              [-81.4008966, 32.11261216],
-              [-81.4027966, 32.10761216],
-              [-81.3998966, 32.10551216],
-              [-81.3962966, 32.10521216],
-              [-81.3915966, 32.10541216],
-              [-81.3889966, 32.10281216],
-              [-81.3917966, 32.09741216],
-              [-81.3914966, 32.09551216],
-              [-81.3883966, 32.09181216],
-              [-81.3882966, 32.08821216],
-              [-81.3875966, 32.08421216],
-              [-81.3864966, 32.08191216],
-              [-81.3846966, 32.07961216],
-              [-81.3793966, 32.07471216],
-              [-81.3755966, 32.07291216],
-              [-81.3722966, 32.07171216],
-              [-81.3695966, 32.06861216],
-              [-81.3713966, 32.06791216],
-              [-81.3739966, 32.06791216],
-              [-81.3747966, 32.06561216],
-              [-81.3720966, 32.06501216],
-              [-81.3701966, 32.06281216],
-              [-81.3689966, 32.05921216],
-              [-81.3672966, 32.05691216],
-              [-81.3695966, 32.05481216],
-              [-81.3681966, 32.05211216],
-              [-81.3671966, 32.04841216],
-              [-81.3638966, 32.04821216],
-              [-81.3611966, 32.04841216],
-              [-81.3612966, 32.04601216],
-              [-81.3562966, 32.04191216],
-              [-81.3542966, 32.03951216],
-              [-81.3525966, 32.03811216],
-              [-81.3512966, 32.03651216],
-              [-81.3502966, 32.03451216],
-              [-81.3488966, 32.03851216],
-              [-81.3475966, 32.03561216],
-              [-81.3433966, 32.03361216],
-              [-81.3409966, 32.03161216],
-              [-81.3382966, 32.03241216],
-              [-81.3356966, 32.02881216],
-              [-81.3350966, 32.02591216],
-              [-81.3362966, 32.02321216],
-              [-81.3345966, 32.02141216],
-              [-81.3316966, 32.02011216],
-              [-81.3278966, 32.01961216],
-              [-81.3246966, 32.02061216],
-              [-81.3236966, 32.01581216],
-              [-81.3197966, 32.01621216],
-              [-81.3187966, 32.01431216],
-              [-81.3205966, 32.01331216],
-              [-81.3220966, 32.01181216],
-              [-81.3225966, 32.00971216],
-              [-81.3199966, 32.00831216],
-              [-81.3177966, 32.01151216],
-              [-81.3148966, 32.01121216],
-              [-81.3146966, 32.00721216],
-              [-81.3134966, 32.00501216],
-              [-81.3113966, 32.00461216],
-              [-81.3082966, 32.00481216],
-              [-81.3058966, 32.00351216],
-              [-81.3041966, 31.99891216],
-              [-81.3092966, 31.99471216],
-              [-81.3072966, 31.99101216],
-              [-81.3072966, 31.98801216],
-              [-81.30828857, 31.98394507],
-              [-81.30828857, 31.98492432],
-              [-81.30780029, 31.98730469],
-              [-81.30938721, 31.98468018],
-              [-81.30938721, 31.98370361],
-              [-81.30889893, 31.98272705],
-              [-81.30718994, 31.98132324],
-              [-81.30249023, 31.97991943],
-              [-81.30090332, 31.97991943],
-              [-81.29968262, 31.98010254],
-              [-81.29382324, 31.98010254],
-              [-81.29052734, 31.97930908],
-              [-81.28912353, 31.97808838],
-              [-81.28851318, 31.97729492],
-              [-81.28747559, 31.9743042],
-              [-81.28771973, 31.97271729],
-              [-81.28869629, 31.96911621],
-              [-81.28887939, 31.96612549],
-              [-81.28918457, 31.96490479],
-              [-81.29022217, 31.96228027],
-              [-81.29040527, 31.96112061],
-              [-81.2902832, 31.95947266],
-              [-81.28961182, 31.95727539],
-              [-81.28851318, 31.95568848],
-              [-81.28710937, 31.95288086],
-              [-81.2869873, 31.95068359],
-              [-81.28710937, 31.95007324],
-              [-81.28879427, 31.9461686],
-              [-81.28839111, 31.94488525],
-              [-81.28778076, 31.9442749],
-              [-81.28637695, 31.94329834],
-              [-81.28540039, 31.94329834],
-              [-81.284729, 31.94372559],
-              [-81.28417969, 31.94348145],
-              [-81.2822876, 31.9442749],
-              [-81.28057861, 31.94549561],
-              [-81.27752685, 31.94488525],
-              [-81.27661133, 31.94451904],
-              [-81.27410889, 31.94268799],
-              [-81.27301025, 31.94152832],
-              [-81.27227783, 31.93890381],
-              [-81.27270508, 31.93707275],
-              [-81.27612305, 31.93188477],
-              [-81.27629207, 31.9303894],
-              [-81.27587891, 31.92828369],
-              [-81.27532959, 31.92730713],
-              [-81.27429199, 31.92608643],
-              [-81.27349853, 31.92572022],
-              [-81.27301025, 31.92529297],
-              [-81.27172852, 31.92468262],
-              [-81.27111816, 31.92468262],
-              [-81.27062988, 31.92492676],
-              [-81.26989746, 31.92449951],
-              [-81.2678833, 31.92468262],
-              [-81.26312256, 31.92730713],
-              [-81.26037598, 31.92791748],
-              [-81.25958252, 31.92791748],
-              [-81.25860596, 31.92749023],
-              [-81.2579956, 31.92687988],
-              [-81.25732422, 31.92468262],
-              [-81.25750732, 31.92370606],
-              [-81.25988769, 31.9216919],
-              [-81.2611084, 31.92089844],
-              [-81.26361084, 31.91967773],
-              [-81.26470947, 31.91851807],
-              [-81.2658081, 31.91589356],
-              [-81.26599121, 31.91430664],
-              [-81.26599121, 31.91229248],
-              [-81.265625, 31.91088867],
-              [-81.26287842, 31.90771484],
-              [-81.2612915, 31.90649414],
-              [-81.26037598, 31.90649414],
-              [-81.25921631, 31.90612793],
-              [-81.25891113, 31.90570068],
-              [-81.25830078, 31.90588379],
-              [-81.25738525, 31.90551758],
-              [-81.25628662, 31.90612793],
-              [-81.25531006, 31.90612793],
-              [-81.25457764, 31.90667725],
-              [-81.25189209, 31.909729],
-              [-81.25140381, 31.91009522],
-              [-81.24981689, 31.91088867],
-              [-81.24609375, 31.91192627],
-              [-81.24468994, 31.91210938],
-              [-81.24090576, 31.91229248],
-              [-81.23779297, 31.91149902],
-              [-81.23522949, 31.91027832],
-              [-81.23309326, 31.9083252],
-              [-81.23291016, 31.90692139],
-              [-81.23308018, 31.90577044],
-              [-81.23272705, 31.90289307],
-              [-81.23358154, 31.90228272],
-              [-81.234375, 31.90228272],
-              [-81.23620605, 31.89990234],
-              [-81.23655325, 31.89984392],
-              [-81.24139404, 31.89630127],
-              [-81.24230957, 31.89532471],
-              [-81.24298096, 31.8939209],
-              [-81.243065, 31.892324],
-              [-81.242362, 31.890731],
-              [-81.237726306383436, 31.885731130383768],
-              [-81.235384187790856, 31.887971710316904],
-              [-81.233618339497397, 31.889811135622598],
-              [-81.231386503459831, 31.891380778550122],
-              [-81.228659274382125, 31.895822828456286],
-              [-81.224766072507521, 31.900052942031579],
-              [-81.220125786772499, 31.906274063242879],
-              [-81.214989626073361, 31.913688943847315],
-              [-81.208949365162638, 31.918784163836118],
-              [-81.203474798841157, 31.920864179465784],
-              [-81.198197299627495, 31.921282517818085],
-              [-81.19446443433003, 31.920574560606497],
-              [-81.191182087258113, 31.918611588337999],
-              [-81.19125541949316, 31.916142613274548],
-              [-81.192944316166958, 31.912146350722463],
-              [-81.195522060755607, 31.909230120315726],
-              [-81.200169973819584, 31.904711315947971],
-              [-81.208174712985311, 31.901741815934876],
-              [-81.214729510471429, 31.901225933799946],
-              [-81.222736888208487, 31.894611143495414],
-              [-81.223742333015977, 31.891039633666484],
-              [-81.219081346198081, 31.88756116961822],
-              [-81.2102966, 31.88621216],
-              [-81.2070966, 31.88731216],
-              [-81.2041966, 31.88941216],
-              [-81.1973966, 31.89151216],
-              [-81.1907966, 31.89441216],
-              [-81.1828966, 31.89721216],
-              [-81.1799966, 31.89761216],
-              [-81.170226055259349, 31.896950702359593],
-              [-81.163404317069237, 31.891204561152694],
-              [-81.159574915896243, 31.886390456820916],
-              [-81.161216087827526, 31.876324602309023],
-              [-81.1641966, 31.86981216],
-              [-81.1620966, 31.86541216],
-              [-81.1584966, 31.86121216],
-              [-81.1558966, 31.85961216],
-              [-81.1478966, 31.85721216],
-              [-81.1443966, 31.85531216],
-              [-81.1428966, 31.85321216],
-              [-81.1446966, 31.85061216],
-              [-81.1471966, 31.84801216],
-              [-81.1488966, 31.84561216],
-              [-81.1519966, 31.84251216],
-              [-81.1531966, 31.84071216],
-              [-81.15342806, 31.84005986],
-              [-81.1541748, 31.83972168],
-              [-81.1552124, 31.83892822],
-              [-81.15661621, 31.83691406],
-              [-81.15759277, 31.83508301],
-              [-81.15838623, 31.83251953],
-              [-81.1585083, 31.82952881],
-              [-81.15917969, 31.82531738],
-              [-81.15991211, 31.82312012],
-              [-81.16088867, 31.82171631],
-              [-81.16137695, 31.82128906],
-              [-81.16351318, 31.82049561],
-              [-81.1651001, 31.82049561],
-              [-81.1663208, 31.8203125],
-              [-81.16729736, 31.82049561],
-              [-81.16888428, 31.8201294],
-              [-81.17108154, 31.8192749],
-              [-81.17462158, 31.81829834],
-              [-81.17559814, 31.81750488],
-              [-81.17608643, 31.81732178],
-              [-81.17712402, 31.81591797],
-              [-81.17712402, 31.81311035],
-              [-81.17669678, 31.80871582],
-              [-81.17681885, 31.80688477],
-              [-81.17718506, 31.8057251],
-              [-81.17962646, 31.80169678],
-              [-81.18078613, 31.79852295],
-              [-81.1809082, 31.7966919],
-              [-81.18121338, 31.79571533],
-              [-81.18109131, 31.79327393],
-              [-81.18041992, 31.79211426],
-              [-81.17919922, 31.79107666],
-              [-81.17730713, 31.79052734],
-              [-81.17657471, 31.79052734],
-              [-81.17602539, 31.7902832],
-              [-81.17510986, 31.7902832],
-              [-81.17132568, 31.78930664],
-              [-81.16870117, 31.78912354],
-              [-81.16802978, 31.7888794],
-              [-81.16729736, 31.7888794],
-              [-81.16552734, 31.78833008],
-              [-81.16491699, 31.78833008],
-              [-81.16308594, 31.78771973],
-              [-81.16192627, 31.78668213],
-              [-81.16082764, 31.78527832],
-              [-81.1605835, 31.78448486],
-              [-81.16040039, 31.78271484],
-              [-81.16137695, 31.77990723],
-              [-81.170556537606103, 31.770058521385216],
-              [-81.171714859911205, 31.767278547853],
-              [-81.172873182216293, 31.764730238781802],
-              [-81.172873182216293, 31.762413594171623],
-              [-81.173220678907811, 31.759865285100421],
-              [-81.172757349985773, 31.75662198264617],
-              [-81.171251530989167, 31.754884499188535],
-              [-81.16847155745694, 31.752683686808862],
-              [-81.161058294704361, 31.7480503975885],
-              [-81.158394153402654, 31.743301276137629],
-              [-81.15862581786368, 31.737162167920651],
-              [-81.1607966, 31.73271216],
-              [-81.1603966, 31.72911216],
-              [-81.1641966, 31.72701216],
-              [-81.1695966, 31.72711216],
-              [-81.1756966, 31.72811216],
-              [-81.1811966, 31.73051216],
-              [-81.1864966, 31.73191216],
-              [-81.1989966, 31.73471216],
-              [-81.2025966, 31.73611216],
-              [-81.2078966, 31.73921216],
-              [-81.2137966, 31.74351216],
-              [-81.2162966, 31.74491216],
-              [-81.2197966, 31.74651216],
-              [-81.2249966, 31.74941216],
-              [-81.2353966, 31.75171216],
-              [-81.2370966, 31.75251216],
-              [-81.2383966, 31.75361216],
-              [-81.2408966, 31.75381216],
-              [-81.2474966, 31.75561216],
-              [-81.2497966, 31.75601216],
-              [-81.2493966, 31.75151216],
-              [-81.2559966, 31.75321216],
-              [-81.2586966, 31.75501216],
-              [-81.2612966, 31.75631216],
-              [-81.2662966, 31.76281216],
-              [-81.2665966, 31.76601216],
-              [-81.2675966, 31.77031216],
-              [-81.2654966, 31.77911216],
-              [-81.2652966, 31.78371216],
-              [-81.2655966, 31.78571216],
-              [-81.2678966, 31.79101216],
-              [-81.2709966, 31.79361216],
-              [-81.2744966, 31.79561216],
-              [-81.2763966, 31.79901216],
-              [-81.2793966, 31.80071216],
-              [-81.2813966, 31.80161216],
-              [-81.2852966, 31.80041216],
-              [-81.289039303332558, 31.800302273190155],
-              [-81.2887966, 31.80231216],
-              [-81.2885966, 31.80591216],
-              [-81.2899966, 31.81011216],
-              [-81.2890966, 31.81201216],
-              [-81.2895966, 31.81471216],
-              [-81.2911966, 31.81751216],
-              [-81.2918966, 31.81441216],
-              [-81.2910966, 31.80771216],
-              [-81.2909966, 31.80451216],
-              [-81.2914966, 31.80181216],
-              [-81.2902966, 31.79931216],
-              [-81.2869966, 31.79801216],
-              [-81.2844966, 31.79861216],
-              [-81.281128046609439, 31.800401577667852],
-              [-81.278777840637304, 31.799805750801674],
-              [-81.2781966, 31.79701216],
-              [-81.2800966, 31.79601216],
-              [-81.28509521, 31.79511261],
-              [-81.2946966, 31.79211216],
-              [-81.306207663780455, 31.790061857551542],
-              [-81.309560286297383, 31.790399228370855],
-              [-81.30807495, 31.79454231],
-              [-81.3072966, 31.79651216],
-              [-81.30749512, 31.80001259],
-              [-81.309787762357871, 31.801986808249126],
-              [-81.3157966, 31.80551216],
-              [-81.3190966, 31.80631216],
-              [-81.3196966, 31.80891216],
-              [-81.3192966, 31.81081216],
-              [-81.3172966, 31.81261216],
-              [-81.3155966, 31.81561216],
-              [-81.3167966, 31.81841216],
-              [-81.3202966, 31.82021216],
-              [-81.3243966, 31.82201216],
-              [-81.3277966, 31.82291216],
-              [-81.3285966, 31.82531216],
-              [-81.3271966, 31.82781216],
-              [-81.3263966, 31.83041216],
-              [-81.3279966, 31.83401216],
-              [-81.3289966, 31.83711216],
-              [-81.3307966, 31.83921216],
-              [-81.3334966, 31.84051216],
-              [-81.3358966, 31.84081216],
-              [-81.3392966, 31.84061216],
-              [-81.3432966, 31.83901216],
-              [-81.3452966, 31.83611216],
-              [-81.3432966, 31.83431216],
-              [-81.3406966, 31.83461216],
-              [-81.3371966, 31.83561216],
-              [-81.3333966, 31.83521216],
-              [-81.3331966, 31.83231216],
-              [-81.3360966, 31.83101216],
-              [-81.3391966, 31.83061216],
-              [-81.3415966, 31.82951216],
-              [-81.3425966, 31.82701216],
-              [-81.3392966, 31.81991216],
-              [-81.3417966, 31.81671216],
-              [-81.3457966, 31.81501216],
-              [-81.3485966, 31.81441216],
-              [-81.3520966, 31.81451216],
-              [-81.3555966, 31.81521216],
-              [-81.3552966, 31.81901216],
-              [-81.3517966, 31.81951216],
-              [-81.3475966, 31.81821216],
-              [-81.3453966, 31.82101216],
-              [-81.3449966, 31.82471216],
-              [-81.3474966, 31.82571216],
-              [-81.3519966, 31.82641216],
-              [-81.3550966, 31.82971216],
-              [-81.3580966, 31.83181216],
-              [-81.3565966, 31.83461216],
-              [-81.3532966, 31.83621216],
-              [-81.3545966, 31.83861216],
-              [-81.3576966, 31.83801216],
-              [-81.3601966, 31.83721216],
-              [-81.3625966, 31.83761216],
-              [-81.3641966, 31.84041216],
-              [-81.3636966, 31.84331216],
-              [-81.3649966, 31.84571216],
-              [-81.3673966, 31.84641216],
-              [-81.3692966, 31.84471216],
-              [-81.3704966, 31.84111216],
-              [-81.3732966, 31.84631216],
-              [-81.3774966, 31.84981216],
-              [-81.3787966, 31.85351216],
-              [-81.3814966, 31.85421216],
-              [-81.3817966, 31.85631216],
-              [-81.3828966, 31.85881216],
-              [-81.3828966, 31.86281216],
-              [-81.3810966, 31.86641216],
-              [-81.3809966, 31.86921216],
-              [-81.3821966, 31.87121216],
-              [-81.3830966, 31.87331216],
-              [-81.3829966, 31.87571216],
-              [-81.3848966, 31.87831216],
-              [-81.3874966, 31.87971216],
-              [-81.3906966, 31.87991216],
-              [-81.3930966, 31.88071216],
-              [-81.3958966, 31.88311216],
-              [-81.3950966, 31.88641216],
-              [-81.3933966, 31.89051216],
-              [-81.3917966, 31.89341216],
-              [-81.3921966, 31.89651216],
-              [-81.3932966, 31.89921216],
-              [-81.3947966, 31.90091216],
-              [-81.3945966, 31.90371216],
-              [-81.3961966, 31.90671216],
-              [-81.3985966, 31.90841216],
-              [-81.4005966, 31.91121216],
-              [-81.4005966, 31.91381216],
-              [-81.4028966, 31.91661216],
-              [-81.4040966, 31.92001216],
-              [-81.4017966, 31.92211216],
-              [-81.4009966, 31.92411216],
-              [-81.4013966, 31.92681216],
-              [-81.4001966, 31.92901216],
-              [-81.4036966, 31.93501216],
-              [-81.4047966, 31.93761216],
-              [-81.4052966, 31.93971216],
-              [-81.4062966, 31.94181216],
-              [-81.4077966, 31.94401216],
-              [-81.4116966, 31.94271216],
-              [-81.4138966, 31.94221216],
-              [-81.4156966, 31.94071216],
-              [-81.4173966, 31.93971216],
-              [-81.4212966, 31.93861216],
-              [-81.4239966, 31.93831216],
-              [-81.4273966, 31.93901216],
-              [-81.4637966, 31.95011216],
-              [-81.4695966, 31.96691216],
-              [-81.4716966, 31.97241216],
-              [-81.4745966, 31.97391216],
-              [-81.4776966, 31.97411216],
-              [-81.4805966, 31.97311216],
-              [-81.4795966, 31.97101216],
-              [-81.4826966, 31.97131216],
-              [-81.4845966, 31.96851216],
-              [-81.4832966, 31.96611216],
-              [-81.4862966, 31.96511216],
-              [-81.4886966, 31.96681216],
-              [-81.4906966, 31.96771216],
-              [-81.4953966, 31.96841216],
-              [-81.4983966, 31.96661216],
-              [-81.5002966, 31.96491216],
-              [-81.5027966, 31.96301216],
-              [-81.5061966, 31.96461216],
-              [-81.5085966, 31.96641216],
-              [-81.5100966, 31.96871216],
-              [-81.5131966, 31.96821216],
-              [-81.5156966, 31.96581216],
-              [-81.5147966, 31.96241216],
-              [-81.5170966, 31.96051216],
-              [-81.5199966, 31.95951216],
-              [-81.5232966, 31.96011216],
-              [-81.5277966, 31.96191216],
-              [-81.5313966, 31.96301216],
-              [-81.5342966, 31.96341216],
-              [-81.5366966, 31.96281216],
-              [-81.5351966, 31.96051216],
-              [-81.5379966, 31.96031216],
-              [-81.5460966, 31.95891216],
-              [-81.5477966, 31.95791216],
-              [-81.5484966, 31.95991216],
-              [-81.5517966, 31.95791216],
-              [-81.5559966, 31.95891216],
-              [-81.5592966, 31.95931216],
-              [-81.5632966, 31.95951216],
-              [-81.5648966, 31.96061216],
-              [-81.5653966, 31.96321216],
-              [-81.5668966, 31.96571216],
-              [-81.5702966, 31.96771216],
-              [-81.5743966, 31.97091216],
-              [-81.5775966, 31.97221216],
-              [-81.5816966, 31.97121216],
-              [-81.5854966, 31.97151216],
-              [-81.5879966, 31.97021216],
-              [-81.5916966, 31.97161216],
-              [-81.5921966, 31.97451216],
-              [-81.5912966, 31.97701216],
-              [-81.5923966, 31.97981216],
-              [-81.5930966, 31.98351216],
-              [-81.5943966, 31.98681216],
-              [-81.5937966, 31.99031216],
-              [-81.6012966, 31.99411216],
-              [-81.5987966, 31.99561216],
-              [-81.6004966, 31.99801216],
-              [-81.6019966, 32.00151216],
-              [-81.6048966, 32.00371216],
-              [-81.6050966, 32.00691216],
-              [-81.6038966, 32.00901216],
-              [-81.6043966, 32.01111216],
-              [-81.6070966, 32.01341216],
-              [-81.6099966, 32.01641216],
-              [-81.6123966, 32.01781216],
-              [-81.6143966, 32.02001216],
-              [-81.6166966, 32.02131216],
-              [-81.6180966, 32.02881216],
-              [-81.6209966, 32.03151216],
-              [-81.6234966, 32.03261216],
-              [-81.6251966, 32.03501216],
-              [-81.6302966, 32.03791216],
-              [-81.6326966, 32.04001216],
-              [-81.6356966, 32.04121216],
-              [-81.6364966, 32.04381216],
-              [-81.6388966, 32.04551216],
-              [-81.6401966, 32.04821216],
-              [-81.6421966, 32.05081216],
-              [-81.6444966, 32.05131216],
-              [-81.6440966, 32.05331216],
-              [-81.6472966, 32.05341216],
-              [-81.6516966, 32.05831216],
-              [-81.6542966, 32.05841216],
-              [-81.6566966, 32.05871216],
-              [-81.6597966, 32.05781216],
-              [-81.6625966, 32.05841216],
-              [-81.6653966, 32.06071216],
-              [-81.6685966, 32.06301216],
-              [-81.6715966, 32.06371216],
-              [-81.6728966, 32.06471216],
-              [-81.6732966, 32.06861216],
-              [-81.6760966, 32.07371216],
-              [-81.6793966, 32.07571216],
-              [-81.6840966, 32.07621216],
-              [-81.6858966, 32.07761216],
-              [-81.6878966, 32.07991216],
-              [-81.6890966, 32.08391216],
-              [-81.6904966, 32.08671216],
-              [-81.6935966, 32.08801216],
-              [-81.6964966, 32.08831216],
-              [-81.6989966, 32.09071216],
-              [-81.7019966, 32.08951216],
-              [-81.7058966, 32.08891216],
-              [-81.7093966, 32.09131216],
-              [-81.7120966, 32.09111216],
-              [-81.7141966, 32.09221216],
-              [-81.7149966, 32.09031216],
-              [-81.71769836, 32.08931273],
-              [-81.72049713, 32.09021378],
-              [-81.72249621, 32.09141192],
-              [-81.7238966, 32.09411216],
-              [-81.72639465, 32.09701157],
-              [-81.72919464, 32.09861374],
-              [-81.73139732, 32.09661304],
-              [-81.73549652, 32.10151291],
-              [-81.74069977, 32.10421372],
-              [-81.7438966, 32.10461216],
-              [-81.74739818, 32.10391239],
-              [-81.75009918, 32.1043129],
-              [-81.7533966, 32.10451216],
-              [-81.7565966, 32.10501216],
-              [-81.75719452, 32.10791397],
-              [-81.7559967, 32.10971069],
-              [-81.75449371, 32.11131287],
-              [-81.7572937, 32.11351395],
-              [-81.7602966, 32.11471216],
-              [-81.76189423, 32.11621094],
-              [-81.76379362, 32.11671138],
-              [-81.7622966, 32.12001216],
-              [-81.7658966, 32.12081216],
-              [-81.76729584, 32.12401199],
-              [-81.7677966, 32.12721216],
-              [-81.76939392, 32.12941361],
-              [-81.77119446, 32.13121414],
-              [-81.7736969, 32.13181305],
-              [-81.77659607, 32.1328125],
-              [-81.77779388, 32.14041138],
-              [-81.78139496, 32.14721298],
-              [-81.78179932, 32.15071106],
-              [-81.78069306, 32.15381238],
-              [-81.7757966, 32.15411216],
-              [-81.6247966, 32.18971216],
-              [-81.5632966, 32.20681216],
-              [-81.4957966, 32.22521216],
-              [-81.4483966, 32.23791216],
-              [-81.4354966, 32.24101216],
-              [-81.4326966, 32.23821216],
-              [-81.4317966, 32.23551216],
-              [-81.4305966, 32.23371216],
-              [-81.4282966, 32.23351216],
-              [-81.4297966, 32.23131216],
-              [-81.4270966, 32.23001216],
-              [-81.4260966, 32.22581216],
-              [-81.4284966, 32.22411216],
-              [-81.4296966, 32.22201216],
-              [-81.4273966, 32.21921216],
-              [-81.4296966, 32.21771216],
-              [-81.4262966, 32.21571216],
-              [-81.4230966, 32.21531216],
-              [-81.4221966, 32.21281216],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        id: "InlandCounties.4",
-        STATE: "GA",
-        CWA: "CHS",
-        COUNTYNAME: "Chatham",
-        FIPS: "13051",
-        TIME_ZONE: "E",
-        FE_AREA: "se",
-        LON: -81.1416,
-        LAT: 32.0099,
-        uuid: "43214e18-2a16-5d8a-a5a4-8db55479e3e2",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-81.239040976469781, 31.888623404573732],
-              [-81.242362, 31.890731],
-              [-81.243065, 31.892324],
-              [-81.24298096, 31.8939209],
-              [-81.24230957, 31.89532471],
-              [-81.24139404, 31.89630127],
-              [-81.23655325, 31.89984392],
-              [-81.23620605, 31.89990234],
-              [-81.234375, 31.90228272],
-              [-81.23358154, 31.90228272],
-              [-81.23272705, 31.90289307],
-              [-81.23308018, 31.90577044],
-              [-81.23291016, 31.90692139],
-              [-81.23309326, 31.9083252],
-              [-81.23522949, 31.91027832],
-              [-81.23779297, 31.91149902],
-              [-81.24090576, 31.91229248],
-              [-81.24468994, 31.91210938],
-              [-81.24609375, 31.91192627],
-              [-81.24981689, 31.91088867],
-              [-81.25140381, 31.91009522],
-              [-81.25189209, 31.909729],
-              [-81.25457764, 31.90667725],
-              [-81.25531006, 31.90612793],
-              [-81.25628662, 31.90612793],
-              [-81.25738525, 31.90551758],
-              [-81.25830078, 31.90588379],
-              [-81.25891113, 31.90570068],
-              [-81.25921631, 31.90612793],
-              [-81.26037598, 31.90649414],
-              [-81.2612915, 31.90649414],
-              [-81.26287842, 31.90771484],
-              [-81.265625, 31.91088867],
-              [-81.26599121, 31.91229248],
-              [-81.26599121, 31.91430664],
-              [-81.2658081, 31.91589356],
-              [-81.26470947, 31.91851807],
-              [-81.26361084, 31.91967773],
-              [-81.2611084, 31.92089844],
-              [-81.25988769, 31.9216919],
-              [-81.25750732, 31.92370606],
-              [-81.25732422, 31.92468262],
-              [-81.2579956, 31.92687988],
-              [-81.25860596, 31.92749023],
-              [-81.25958252, 31.92791748],
-              [-81.26037598, 31.92791748],
-              [-81.26312256, 31.92730713],
-              [-81.2678833, 31.92468262],
-              [-81.26989746, 31.92449951],
-              [-81.27062988, 31.92492676],
-              [-81.27111816, 31.92468262],
-              [-81.27172852, 31.92468262],
-              [-81.27301025, 31.92529297],
-              [-81.27349853, 31.92572022],
-              [-81.27429199, 31.92608643],
-              [-81.27532959, 31.92730713],
-              [-81.27587891, 31.92828369],
-              [-81.27629207, 31.9303894],
-              [-81.27612305, 31.93188477],
-              [-81.27270508, 31.93707275],
-              [-81.27227783, 31.93890381],
-              [-81.27301025, 31.94152832],
-              [-81.27410889, 31.94268799],
-              [-81.27661133, 31.94451904],
-              [-81.27752685, 31.94488525],
-              [-81.28057861, 31.94549561],
-              [-81.2822876, 31.9442749],
-              [-81.28417969, 31.94348145],
-              [-81.284729, 31.94372559],
-              [-81.28540039, 31.94329834],
-              [-81.28637695, 31.94329834],
-              [-81.28778076, 31.9442749],
-              [-81.28839111, 31.94488525],
-              [-81.28879427, 31.9461686],
-              [-81.28710937, 31.95007324],
-              [-81.2869873, 31.95068359],
-              [-81.28710937, 31.95288086],
-              [-81.28851318, 31.95568848],
-              [-81.28961182, 31.95727539],
-              [-81.2902832, 31.95947266],
-              [-81.29040527, 31.96112061],
-              [-81.29022217, 31.96228027],
-              [-81.28918457, 31.96490479],
-              [-81.28887939, 31.96612549],
-              [-81.28869629, 31.96911621],
-              [-81.28771973, 31.97271729],
-              [-81.28747559, 31.9743042],
-              [-81.28851318, 31.97729492],
-              [-81.28912353, 31.97808838],
-              [-81.29052734, 31.97930908],
-              [-81.29382324, 31.98010254],
-              [-81.29968262, 31.98010254],
-              [-81.30090332, 31.97991943],
-              [-81.30249023, 31.97991943],
-              [-81.30718994, 31.98132324],
-              [-81.30889893, 31.98272705],
-              [-81.30938721, 31.98370361],
-              [-81.30938721, 31.98468018],
-              [-81.30780029, 31.98730469],
-              [-81.30828857, 31.98492432],
-              [-81.30828857, 31.98394507],
-              [-81.3072966, 31.98801216],
-              [-81.3072966, 31.99101216],
-              [-81.3092966, 31.99471216],
-              [-81.3041966, 31.99891216],
-              [-81.3058966, 32.00351216],
-              [-81.3082966, 32.00481216],
-              [-81.3113966, 32.00461216],
-              [-81.3134966, 32.00501216],
-              [-81.3146966, 32.00721216],
-              [-81.3148966, 32.01121216],
-              [-81.3177966, 32.01151216],
-              [-81.3199966, 32.00831216],
-              [-81.3225966, 32.00971216],
-              [-81.3220966, 32.01181216],
-              [-81.3205966, 32.01331216],
-              [-81.3187966, 32.01431216],
-              [-81.3197966, 32.01621216],
-              [-81.3236966, 32.01581216],
-              [-81.3246966, 32.02061216],
-              [-81.3278966, 32.01961216],
-              [-81.3316966, 32.02011216],
-              [-81.3345966, 32.02141216],
-              [-81.3362966, 32.02321216],
-              [-81.3350966, 32.02591216],
-              [-81.3356966, 32.02881216],
-              [-81.3382966, 32.03241216],
-              [-81.3409966, 32.03161216],
-              [-81.3433966, 32.03361216],
-              [-81.3475966, 32.03561216],
-              [-81.3488966, 32.03851216],
-              [-81.3502966, 32.03451216],
-              [-81.3512966, 32.03651216],
-              [-81.3525966, 32.03811216],
-              [-81.3542966, 32.03951216],
-              [-81.3562966, 32.04191216],
-              [-81.3612966, 32.04601216],
-              [-81.3611966, 32.04841216],
-              [-81.3638966, 32.04821216],
-              [-81.3671966, 32.04841216],
-              [-81.3681966, 32.05211216],
-              [-81.3695966, 32.05481216],
-              [-81.3672966, 32.05691216],
-              [-81.3689966, 32.05921216],
-              [-81.3701966, 32.06281216],
-              [-81.3720966, 32.06501216],
-              [-81.3747966, 32.06561216],
-              [-81.3739966, 32.06791216],
-              [-81.3713966, 32.06791216],
-              [-81.3695966, 32.06861216],
-              [-81.3722966, 32.07171216],
-              [-81.3755966, 32.07291216],
-              [-81.3793966, 32.07471216],
-              [-81.3846966, 32.07961216],
-              [-81.3864966, 32.08191216],
-              [-81.3875966, 32.08421216],
-              [-81.3882966, 32.08821216],
-              [-81.3883966, 32.09181216],
-              [-81.3914966, 32.09551216],
-              [-81.38751221, 32.09912109],
-              [-81.38488769, 32.10107422],
-              [-81.34002685, 32.13348389],
-              [-81.33187252, 32.13894422],
-              [-81.32818603, 32.14190674],
-              [-81.31561279, 32.15087891],
-              [-81.31262207, 32.15332031],
-              [-81.30981445, 32.15509033],
-              [-81.30450439, 32.15887451],
-              [-81.27069092, 32.18347168],
-              [-81.25299072, 32.19610596],
-              [-81.23931885, 32.20629883],
-              [-81.22918701, 32.21331787],
-              [-81.22412109, 32.21710205],
-              [-81.2164917, 32.22247315],
-              [-81.20141602, 32.23370361],
-              [-81.19544317, 32.23758596],
-              [-81.1892966, 32.23641216],
-              [-81.145434, 32.226699],
-              [-81.113107430666972, 32.195625137154813],
-              [-81.108733550866418, 32.172100756605907],
-              [-81.117026833315862, 32.168312549402991],
-              [-81.108630968178389, 32.144817236032665],
-              [-81.113256194574006, 32.126672117096028],
-              [-81.11675668508822, 32.117075760289673],
-              [-81.106497614912556, 32.111813727985542],
-              [-81.097601373247301, 32.11173801529052],
-              [-81.092377197290688, 32.111624446247987],
-              [-81.090597948957637, 32.109958766957469],
-              [-81.090295098177549, 32.107346678979162],
-              [-81.084692358745826, 32.102008933980017],
-              [-81.079729985709164, 32.09819974715132],
-              [-81.055507715249561, 32.086075914524805],
-              [-81.03601074, 32.08251953],
-              [-81.02990723, 32.08392334],
-              [-81.026923410171804, 32.08733130630187],
-              [-81.006773994106609, 32.102010446047061],
-              [-80.988436091495686, 32.084980611215244],
-              [-80.976624831642937, 32.075700335616638],
-              [-80.953958867156246, 32.067330884887035],
-              [-80.927764834478864, 32.042882339995828],
-              [-80.91419108593152, 32.037092295035769],
-              [-80.878252675140885, 32.03217781950714],
-              [-80.856753071860538, 32.006901258893755],
-              [-80.868083943859645, 31.987725937049117],
-              [-80.917474924368562, 31.985692190792868],
-              [-80.923188050105793, 31.956490575158597],
-              [-80.961834985531212, 31.940083857289313],
-              [-80.993119622447708, 31.90822768529047],
-              [-80.988261932475709, 31.875843085477133],
-              [-81.004656636131216, 31.864710879291295],
-              [-81.044327770902555, 31.863901264295961],
-              [-81.062948915795232, 31.842446466919625],
-              [-81.090678229385404, 31.850542616872961],
-              [-81.09527588, 31.85827637],
-              [-81.09729004, 31.85870361],
-              [-81.09991455, 31.86090088],
-              [-81.10137939, 31.8626709],
-              [-81.10308838, 31.86431885],
-              [-81.10681152, 31.86608887],
-              [-81.1072998, 31.86688232],
-              [-81.109375, 31.86688232],
-              [-81.11029053, 31.86651611],
-              [-81.11242676, 31.86608887],
-              [-81.11328125, 31.86572266],
-              [-81.1149292, 31.86468506],
-              [-81.11602783, 31.86431885],
-              [-81.11730957, 31.86328125],
-              [-81.12249756, 31.86047363],
-              [-81.1262207, 31.85968018],
-              [-81.12939453, 31.85949707],
-              [-81.13220215, 31.85913086],
-              [-81.13812256, 31.85968018],
-              [-81.14099121, 31.86047363],
-              [-81.14270019, 31.86071777],
-              [-81.14752197, 31.86248779],
-              [-81.14880371, 31.8638916],
-              [-81.15307617, 31.86950684],
-              [-81.15307617, 31.87011719],
-              [-81.15252685, 31.87011719],
-              [-81.15258789, 31.87091065],
-              [-81.15350342, 31.87292481],
-              [-81.15588379, 31.8770752],
-              [-81.15612793, 31.87811279],
-              [-81.15588379, 31.88110352],
-              [-81.15588379, 31.88452148],
-              [-81.15631103, 31.88690186],
-              [-81.15838623, 31.89227295],
-              [-81.16461182, 31.89831543],
-              [-81.16748047, 31.90032959],
-              [-81.1730957, 31.90307617],
-              [-81.17431641, 31.90429688],
-              [-81.17779541, 31.90447998],
-              [-81.18157959, 31.90527344],
-              [-81.18450928, 31.90509033],
-              [-81.199773320691378, 31.896973011383476],
-              [-81.204686813509369, 31.892300987560496],
-              [-81.211208754432178, 31.888309109926706],
-              [-81.21703128504106, 31.887955881899799],
-              [-81.220309641605184, 31.889539806981347],
-              [-81.223088795042514, 31.891864564755981],
-              [-81.21949926133091, 31.8954703171704],
-              [-81.205143738940791, 31.89880681273694],
-              [-81.193486001765223, 31.908106805315199],
-              [-81.190080370680221, 31.919502570868847],
-              [-81.194402902441951, 31.921467358033269],
-              [-81.202262051099638, 31.922253272899034],
-              [-81.210900430631511, 31.918875248319466],
-              [-81.220152184370193, 31.909906025991948],
-              [-81.228635487946235, 31.898212037895608],
-              [-81.231833197879112, 31.89140279674443],
-              [-81.236761433187411, 31.887302204006982],
-              [-81.239040976469781, 31.888623404573732],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        id: "InlandCounties.5",
-        STATE: "GA",
-        CWA: "JAX",
-        COUNTYNAME: "Glynn",
-        FIPS: "13127",
-        TIME_ZONE: "E",
-        FE_AREA: "se",
-        LON: -81.5512,
-        LAT: 31.2365,
-        uuid: "d6375cd7-a126-534d-a661-951fc482ce63",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-81.56232452, 31.12861633],
-              [-81.56256216, 31.12843364],
-              [-81.56584, 31.129487],
-              [-81.566107, 31.129716],
-              [-81.566053, 31.129945],
-              [-81.565148, 31.129968],
-              [-81.56416321, 31.12971497],
-              [-81.56232452, 31.12884522],
-              [-81.56232452, 31.12861633],
-            ],
-          ],
-          [
-            [
-              [-81.5803966, 31.40141216],
-              [-81.5776966, 31.39871216],
-              [-81.5730966, 31.39501216],
-              [-81.5703966, 31.39321216],
-              [-81.5657966, 31.39111216],
-              [-81.5630966, 31.38891216],
-              [-81.5584966, 31.38611216],
-              [-81.5512966, 31.38591216],
-              [-81.5469966, 31.38371216],
-              [-81.5486966, 31.38051216],
-              [-81.5520966, 31.37891216],
-              [-81.5520966, 31.37621216],
-              [-81.54987324, 31.375224],
-              [-81.550307, 31.375224],
-              [-81.549925, 31.374934],
-              [-81.549448, 31.374834],
-              [-81.548868, 31.374966],
-              [-81.548598, 31.375224],
-              [-81.54889144, 31.375224],
-              [-81.5462966, 31.37631216],
-              [-81.5422966, 31.37691216],
-              [-81.5394966, 31.37611216],
-              [-81.5386966, 31.37301216],
-              [-81.5375966, 31.37031216],
-              [-81.5358966, 31.36761216],
-              [-81.5306966, 31.36621216],
-              [-81.5279966, 31.36511216],
-              [-81.5254966, 31.36341216],
-              [-81.5230966, 31.36031216],
-              [-81.5209966, 31.35721216],
-              [-81.5191966, 31.35641216],
-              [-81.5175966, 31.35541216],
-              [-81.5089966, 31.35491216],
-              [-81.5056966, 31.35421216],
-              [-81.5026966, 31.35251216],
-              [-81.4966966, 31.34741216],
-              [-81.4929966, 31.34501216],
-              [-81.49012999, 31.343456],
-              [-81.49005788, 31.34423906],
-              [-81.48934, 31.344654],
-              [-81.489153, 31.343692],
-              [-81.48918152, 31.3420887],
-              [-81.48870087, 31.34206581],
-              [-81.486082, 31.34289101],
-              [-81.48514557, 31.34257126],
-              [-81.48471832, 31.34220314],
-              [-81.484479, 31.341517],
-              [-81.48447418, 31.33998108],
-              [-81.48431428, 31.33984592],
-              [-81.482182, 31.340212],
-              [-81.4815979, 31.34000588],
-              [-81.48140717, 31.33966255],
-              [-81.48255157, 31.33787346],
-              [-81.48260498, 31.33748436],
-              [-81.48226882, 31.33718346],
-              [-81.48065948, 31.33745194],
-              [-81.48028564, 31.33789825],
-              [-81.479885, 31.337944],
-              [-81.47926976, 31.33768483],
-              [-81.4784966, 31.33781216],
-              [-81.47826385, 31.33795929],
-              [-81.47801971, 31.33867836],
-              [-81.47665405, 31.33964157],
-              [-81.47638728, 31.33941482],
-              [-81.475209, 31.339634],
-              [-81.47496033, 31.33911896],
-              [-81.474914, 31.336639],
-              [-81.474446, 31.334612],
-              [-81.47387695, 31.33352661],
-              [-81.473119, 31.332949],
-              [-81.4723053, 31.33285523],
-              [-81.47134399, 31.33301163],
-              [-81.47073486, 31.33341518],
-              [-81.47035449, 31.33390347],
-              [-81.470562, 31.3345071],
-              [-81.472184, 31.3371511],
-              [-81.47248077, 31.33833695],
-              [-81.472192, 31.338999],
-              [-81.471652, 31.339508],
-              [-81.47070312, 31.3398571],
-              [-81.46974182, 31.33982849],
-              [-81.4689, 31.339404],
-              [-81.468177, 31.338794],
-              [-81.467649, 31.338014],
-              [-81.46700287, 31.33657455],
-              [-81.466423, 31.336045],
-              [-81.463864, 31.335063],
-              [-81.461020354047761, 31.332563659228615],
-              [-81.458236320000978, 31.330423031973652],
-              [-81.45665741, 31.32860375],
-              [-81.454525, 31.326781],
-              [-81.453606, 31.326502],
-              [-81.452241, 31.326691],
-              [-81.447279, 31.328397],
-              [-81.44107056, 31.33195877],
-              [-81.439706, 31.332259],
-              [-81.43793487, 31.33203125],
-              [-81.4366684, 31.33148575],
-              [-81.43560028, 31.33074188],
-              [-81.43529434, 31.32884413],
-              [-81.435902, 31.3242541],
-              [-81.437527, 31.3191381],
-              [-81.43782806, 31.31757355],
-              [-81.43753815, 31.31598473],
-              [-81.43689728, 31.31456184],
-              [-81.43596649, 31.31344414],
-              [-81.43495178, 31.31253815],
-              [-81.43338776, 31.31215668],
-              [-81.42933655, 31.31186295],
-              [-81.426762, 31.311386],
-              [-81.423116, 31.309834],
-              [-81.41799728, 31.30679942],
-              [-81.414355599175082, 31.305629918412773],
-              [-81.410676639135048, 31.307933348578064],
-              [-81.40893151431797, 31.309174326225765],
-              [-81.406294436816609, 31.309445790086198],
-              [-81.404743214756976, 31.306808712584836],
-              [-81.404006384278659, 31.305451393282663],
-              [-81.404935143360817, 31.304081243356364],
-              [-81.408365409965683, 31.302164101571087],
-              [-81.413445662210947, 31.295648968920663],
-              [-81.412266641746157, 31.293429422320571],
-              [-81.411258347407411, 31.291916980812438],
-              [-81.409474442038842, 31.289822831031941],
-              [-81.407966, 31.288332],
-              [-81.406659, 31.287897],
-              [-81.404504422898654, 31.288251089634596],
-              [-81.400708, 31.288285],
-              [-81.398386, 31.288101],
-              [-81.393876, 31.285991],
-              [-81.391101, 31.284913],
-              [-81.389421, 31.283697],
-              [-81.388194, 31.28207],
-              [-81.387741, 31.281313],
-              [-81.387608, 31.280075],
-              [-81.387368, 31.279961],
-              [-81.38684025, 31.27823617],
-              [-81.3867569, 31.27677536],
-              [-81.38689, 31.276225],
-              [-81.38745117, 31.27457619],
-              [-81.38787842, 31.27409554],
-              [-81.377412342942506, 31.255140031143686],
-              [-81.382623591973982, 31.232178268227578],
-              [-81.39509058495635, 31.20510547090991],
-              [-81.407902126486761, 31.172109419058945],
-              [-81.42803529409322, 31.125256306789584],
-              [-81.430853961746053, 31.103359695189127],
-              [-81.423897310682989, 31.062398829597171],
-              [-81.425887211986279, 31.034635520005306],
-              [-81.48051468005103, 31.032470520741899],
-              [-81.487626190021061, 31.041570233660366],
-              [-81.490575483309911, 31.054210062041111],
-              [-81.501530001239885, 31.056906558762339],
-              [-81.512516540564917, 31.059419615252271],
-              [-81.528157906361997, 31.064574721313324],
-              [-81.533050813311021, 31.076504730160085],
-              [-81.53296473403708, 31.078232049555808],
-              [-81.538126752477453, 31.08382423619955],
-              [-81.546113299117266, 31.0829875512492],
-              [-81.553681439015179, 31.080578117235437],
-              [-81.558452200086435, 31.081780937233571],
-              [-81.560925667255788, 31.086835413623106],
-              [-81.56339913442514, 31.091352179758438],
-              [-81.571909181624491, 31.093666131510339],
-              [-81.571838724385032, 31.099020881708626],
-              [-81.568386319651935, 31.101768714047221],
-              [-81.562820197735292, 31.106771178048252],
-              [-81.560847395030663, 31.115789704697999],
-              [-81.55760956, 31.12095833],
-              [-81.55737, 31.122471],
-              [-81.557889, 31.12523],
-              [-81.55789947, 31.12705803],
-              [-81.55821991, 31.12763023],
-              [-81.55888367, 31.1282959],
-              [-81.56211, 31.130449],
-              [-81.56482696, 31.1312294],
-              [-81.566373, 31.131137],
-              [-81.569678, 31.130082],
-              [-81.57202148, 31.12873077],
-              [-81.574073, 31.127034],
-              [-81.5756073, 31.12523079],
-              [-81.57607269, 31.12409782],
-              [-81.57666016, 31.12086678],
-              [-81.57700348, 31.11976624],
-              [-81.577563, 31.118483],
-              [-81.578255, 31.117405],
-              [-81.578761, 31.116878],
-              [-81.5813446, 31.11573219],
-              [-81.582383, 31.115502],
-              [-81.58288574, 31.11566353],
-              [-81.58347321, 31.11593819],
-              [-81.58584595, 31.11774826],
-              [-81.587126, 31.118435],
-              [-81.58959961, 31.11905289],
-              [-81.59021759, 31.11900711],
-              [-81.59085846, 31.1186409],
-              [-81.59429169, 31.11625671],
-              [-81.59477234, 31.11607361],
-              [-81.59767151, 31.11641693],
-              [-81.59928402, 31.11634636],
-              [-81.6014328, 31.11584282],
-              [-81.60800934, 31.11336327],
-              [-81.60955048, 31.11315918],
-              [-81.61048889, 31.11334038],
-              [-81.61395264, 31.11469078],
-              [-81.61470032, 31.1152401],
-              [-81.615448, 31.11615753],
-              [-81.61560821, 31.11663818],
-              [-81.61693573, 31.11750984],
-              [-81.61871338, 31.11769104],
-              [-81.62154388, 31.11739159],
-              [-81.62380981, 31.11649513],
-              [-81.62545013, 31.11541939],
-              [-81.62625122, 31.11475563],
-              [-81.62828064, 31.11177826],
-              [-81.63008881, 31.1099453],
-              [-81.63176727, 31.10862923],
-              [-81.63301849, 31.10829544],
-              [-81.63362884, 31.10873222],
-              [-81.63471985, 31.11038208],
-              [-81.63549805, 31.11125374],
-              [-81.64073944, 31.11574745],
-              [-81.64159393, 31.11758041],
-              [-81.64212036, 31.11815453],
-              [-81.64266968, 31.11831284],
-              [-81.64373016, 31.11799431],
-              [-81.6461792, 31.11581802],
-              [-81.64654541, 31.11558723],
-              [-81.64723968, 31.11558723],
-              [-81.64899445, 31.11701012],
-              [-81.64990234, 31.11820221],
-              [-81.65070343, 31.11966896],
-              [-81.65163422, 31.12072372],
-              [-81.65219116, 31.12104416],
-              [-81.65331268, 31.12129784],
-              [-81.65496826, 31.1213665],
-              [-81.65756226, 31.12205124],
-              [-81.6574173, 31.12250328],
-              [-81.6579361, 31.12235832],
-              [-81.6588974, 31.12401199],
-              [-81.66189575, 31.12511253],
-              [-81.66381836, 31.12521172],
-              [-81.68610382, 31.12231636],
-              [-81.69109344, 31.12571144],
-              [-81.69389343, 31.1274128],
-              [-81.70329285, 31.13421249],
-              [-81.71029663, 31.13671303],
-              [-81.71259308, 31.13851166],
-              [-81.7151947, 31.14171219],
-              [-81.71649933, 31.14451218],
-              [-81.72189331, 31.14861298],
-              [-81.72409821, 31.14951134],
-              [-81.72879791, 31.14961243],
-              [-81.73199463, 31.15071297],
-              [-81.73580933, 31.15161133],
-              [-81.74032593, 31.15151215],
-              [-81.74443054, 31.1511116],
-              [-81.74723053, 31.15101242],
-              [-81.7508316, 31.15051269],
-              [-81.75429535, 31.15051269],
-              [-81.76269531, 31.15371323],
-              [-81.76439667, 31.15651131],
-              [-81.76439667, 31.16001129],
-              [-81.76309967, 31.16341209],
-              [-81.76309967, 31.16671181],
-              [-81.76519775, 31.16921234],
-              [-81.76832581, 31.16901207],
-              [-81.77302551, 31.16531181],
-              [-81.7763977, 31.16471291],
-              [-81.7796936, 31.16501236],
-              [-81.7818985, 31.16641235],
-              [-81.78129578, 31.1729126],
-              [-81.7763977, 31.17631149],
-              [-81.77379608, 31.18261147],
-              [-81.76669311, 31.20181274],
-              [-81.76469421, 31.22141266],
-              [-81.76359558, 31.22471237],
-              [-81.76210022, 31.22681236],
-              [-81.75889587, 31.23681259],
-              [-81.7579956, 31.24011231],
-              [-81.75709534, 31.24541283],
-              [-81.751297, 31.268713],
-              [-81.75059509, 31.27131271],
-              [-81.74720001, 31.28141213],
-              [-81.74459839, 31.28391266],
-              [-81.73779297, 31.28901291],
-              [-81.73649597, 31.29101181],
-              [-81.73449707, 31.29501152],
-              [-81.7335968, 31.29771233],
-              [-81.7328949, 31.30021286],
-              [-81.73269653, 31.30431175],
-              [-81.73299408, 31.32131195],
-              [-81.7328949, 31.3257122],
-              [-81.7315979, 31.33001137],
-              [-81.72439575, 31.32791138],
-              [-81.70769501, 31.34831238],
-              [-81.69889831, 31.36291122],
-              [-81.69799805, 31.36781311],
-              [-81.70459747, 31.37581253],
-              [-81.70219421, 31.38221169],
-              [-81.69919586, 31.38861275],
-              [-81.69799805, 31.3909111],
-              [-81.69400024, 31.39721298],
-              [-81.69219971, 31.40041161],
-              [-81.68719482, 31.40041161],
-              [-81.66539764, 31.3998127],
-              [-81.66209412, 31.40371132],
-              [-81.62869263, 31.44911194],
-              [-81.625, 31.45311165],
-              [-81.6249966, 31.45001216],
-              [-81.6207966, 31.44741216],
-              [-81.6188966, 31.44631216],
-              [-81.6162966, 31.44571216],
-              [-81.6121966, 31.44561216],
-              [-81.6112966, 31.44321216],
-              [-81.6099966, 31.44141216],
-              [-81.6077966, 31.43931216],
-              [-81.61049659, 31.43811216],
-              [-81.6120966, 31.43571216],
-              [-81.6107966, 31.43391216],
-              [-81.6076966, 31.43221216],
-              [-81.6059966, 31.43051216],
-              [-81.6062966, 31.42831216],
-              [-81.6059966, 31.42631216],
-              [-81.6034966, 31.42381216],
-              [-81.5988966, 31.42131216],
-              [-81.5961966, 31.41771216],
-              [-81.5951966, 31.41501216],
-              [-81.5927966, 31.41491216],
-              [-81.5905966, 31.41321216],
-              [-81.5887966, 31.40991216],
-              [-81.5852966, 31.40901216],
-              [-81.5829966, 31.40741216],
-              [-81.5810966, 31.40511216],
-              [-81.5803966, 31.40141216],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        id: "InlandCounties.6",
-        STATE: "GA",
-        CWA: "CHS",
-        COUNTYNAME: "McIntosh",
-        FIPS: "13191",
-        TIME_ZONE: "E",
-        FE_AREA: "se",
-        LON: -81.4149,
-        LAT: 31.5007,
-        uuid: "4ca467df-916d-5a9e-86a4-263fc35b2a7b",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-81.48633, 31.699515],
-              [-81.48493, 31.699315],
-              [-81.48243, 31.699316],
-              [-81.47663, 31.698516],
-              [-81.47453, 31.697816],
-              [-81.47063, 31.695816],
-              [-81.46873, 31.695216],
-              [-81.465629, 31.693616],
-              [-81.464929, 31.692416],
-              [-81.464629, 31.690916],
-              [-81.464229, 31.690416],
-              [-81.463529, 31.690016],
-              [-81.462329, 31.688916],
-              [-81.461129, 31.687216],
-              [-81.461263, 31.685211],
-              [-81.46104, 31.684044],
-              [-81.461101, 31.683215],
-              [-81.461839, 31.682349],
-              [-81.464529, 31.681316],
-              [-81.466529, 31.680916],
-              [-81.467129, 31.680616],
-              [-81.468029, 31.679616],
-              [-81.467929, 31.678916],
-              [-81.465629, 31.676316],
-              [-81.462329, 31.673316],
-              [-81.461529, 31.672416],
-              [-81.457529, 31.662817],
-              [-81.456629, 31.661517],
-              [-81.456629, 31.661317],
-              [-81.452928, 31.658417],
-              [-81.451928, 31.657817],
-              [-81.448728, 31.656817],
-              [-81.445728, 31.656217],
-              [-81.443128, 31.654917],
-              [-81.442028, 31.653817],
-              [-81.441828, 31.653417],
-              [-81.439528, 31.647117],
-              [-81.438528, 31.645517],
-              [-81.437728, 31.643617],
-              [-81.436928, 31.642917],
-              [-81.434927, 31.641717],
-              [-81.434527, 31.641317],
-              [-81.432127, 31.640518],
-              [-81.431127, 31.640518],
-              [-81.429527, 31.641018],
-              [-81.428127, 31.642517],
-              [-81.427827, 31.642617],
-              [-81.426927, 31.643717],
-              [-81.426627, 31.644717],
-              [-81.426642, 31.64529],
-              [-81.4246966, 31.64901216],
-              [-81.4212966, 31.65041216],
-              [-81.4162966, 31.65211216],
-              [-81.4142966, 31.65001216],
-              [-81.4116966, 31.64881216],
-              [-81.4099966, 31.64751216],
-              [-81.4072966, 31.64841216],
-              [-81.4040966, 31.64731216],
-              [-81.4010966, 31.64571216],
-              [-81.3963966, 31.64351216],
-              [-81.3931186, 31.64359621],
-              [-81.392779, 31.643473],
-              [-81.391241, 31.643629],
-              [-81.390652, 31.64406],
-              [-81.39044, 31.644533],
-              [-81.390578, 31.644768],
-              [-81.391647, 31.645243],
-              [-81.390732, 31.646228],
-              [-81.390111, 31.64528],
-              [-81.389481, 31.645163],
-              [-81.389289, 31.645484],
-              [-81.389318, 31.646185],
-              [-81.389176, 31.646248],
-              [-81.388559, 31.645873],
-              [-81.388059, 31.645869],
-              [-81.387734, 31.646294],
-              [-81.387701, 31.64668],
-              [-81.386814, 31.647632],
-              [-81.386871, 31.64843],
-              [-81.387218, 31.648634],
-              [-81.386906, 31.648664],
-              [-81.386309, 31.648152],
-              [-81.384988, 31.646547],
-              [-81.383911, 31.646495],
-              [-81.382705, 31.647251],
-              [-81.382196, 31.648488],
-              [-81.381954, 31.648557],
-              [-81.381097, 31.648121],
-              [-81.380321, 31.648189],
-              [-81.379972, 31.648395],
-              [-81.379838, 31.64945],
-              [-81.380105, 31.650045],
-              [-81.380051, 31.65055],
-              [-81.37989, 31.650664],
-              [-81.377855, 31.649655],
-              [-81.377105, 31.6497],
-              [-81.37673, 31.649884],
-              [-81.376408, 31.650204],
-              [-81.376354, 31.651442],
-              [-81.375898, 31.652496],
-              [-81.375897, 31.653229],
-              [-81.376459, 31.653894],
-              [-81.377584, 31.654238],
-              [-81.377638, 31.654697],
-              [-81.377477, 31.654765],
-              [-81.376245, 31.654467],
-              [-81.375277, 31.653733],
-              [-81.3752765, 31.6522315],
-              [-81.374427, 31.651349],
-              [-81.373457, 31.650708],
-              [-81.373243, 31.650776],
-              [-81.373109, 31.650639],
-              [-81.373055, 31.650341],
-              [-81.373617, 31.649608],
-              [-81.374126, 31.649355],
-              [-81.374825, 31.649424],
-              [-81.375738, 31.648829],
-              [-81.375953, 31.647935],
-              [-81.375418, 31.647247],
-              [-81.374825, 31.646938],
-              [-81.373535, 31.646949],
-              [-81.371152, 31.647844],
-              [-81.37075, 31.648601],
-              [-81.37126, 31.65025],
-              [-81.371261, 31.651121],
-              [-81.370565, 31.651694],
-              [-81.370082, 31.65158],
-              [-81.36968, 31.651191],
-              [-81.368501, 31.649266],
-              [-81.368098, 31.648923],
-              [-81.36692, 31.648442],
-              [-81.366196, 31.648465],
-              [-81.365393, 31.648649],
-              [-81.364322, 31.650139],
-              [-81.364162, 31.65062],
-              [-81.364136, 31.652385],
-              [-81.364297, 31.653576],
-              [-81.363654, 31.654081],
-              [-81.363226, 31.654127],
-              [-81.36261, 31.653944],
-              [-81.362154, 31.653623],
-              [-81.361484, 31.652386],
-              [-81.360465, 31.651836],
-              [-81.35851, 31.652249],
-              [-81.357519, 31.652708],
-              [-81.356126, 31.653786],
-              [-81.355242, 31.654084],
-              [-81.354144, 31.65358],
-              [-81.352027, 31.653489],
-              [-81.350259, 31.653054],
-              [-81.347312, 31.652184],
-              [-81.346428, 31.651772],
-              [-81.345676, 31.650901],
-              [-81.34506, 31.649251],
-              [-81.344657, 31.64861],
-              [-81.344068, 31.648358],
-              [-81.343398, 31.647785],
-              [-81.340183, 31.646388],
-              [-81.33871, 31.646251],
-              [-81.336299, 31.646985],
-              [-81.335897, 31.647374],
-              [-81.335522, 31.648016],
-              [-81.334907, 31.649918],
-              [-81.334505, 31.650491],
-              [-81.33046, 31.651454],
-              [-81.327192, 31.652417],
-              [-81.325344, 31.653174],
-              [-81.32454, 31.653655],
-              [-81.322397, 31.655535],
-              [-81.321084, 31.657162],
-              [-81.319825, 31.657987],
-              [-81.31795, 31.658445],
-              [-81.316209, 31.658285],
-              [-81.313796, 31.657414],
-              [-81.31252911, 31.65634737],
-              [-81.311707, 31.655031],
-              [-81.307822, 31.651295],
-              [-81.303456, 31.649394],
-              [-81.298929, 31.647766],
-              [-81.27065020704913, 31.651924322691197],
-              [-81.261328438803034, 31.648030893383588],
-              [-81.220115749840275, 31.628011147915498],
-              [-81.215385734179691, 31.62001312143488],
-              [-81.209537714817515, 31.61321909894059],
-              [-81.203704, 31.6075],
-              [-81.195541943726809, 31.600539648576703],
-              [-81.197714727075436, 31.558170373278308],
-              [-81.253806908454933, 31.539403811356873],
-              [-81.258897, 31.528975],
-              [-81.258809, 31.52906],
-              [-81.253223271117562, 31.503288852697114],
-              [-81.252845779065296, 31.305816761164593],
-              [-81.278040713837299, 31.302176801460853],
-              [-81.295408584023292, 31.291290720148105],
-              [-81.302804471327363, 31.296692210875804],
-              [-81.305962265906643, 31.302509200890249],
-              [-81.3174626203879, 31.30913444936018],
-              [-81.327893712063499, 31.307964470349138],
-              [-81.330525431457446, 31.306256899194373],
-              [-81.331764108253481, 31.306004776075724],
-              [-81.333870779597405, 31.306984623212433],
-              [-81.333527833099552, 31.310316103477234],
-              [-81.356539687173637, 31.314303740266887],
-              [-81.367427505242333, 31.320717112553925],
-              [-81.380403398009136, 31.31937477881943],
-              [-81.400502984447002, 31.300997388001136],
-              [-81.403347114903326, 31.295252108240629],
-              [-81.407730615681032, 31.292574809109919],
-              [-81.411436990749891, 31.293152426003765],
-              [-81.411822068679115, 31.296473723143389],
-              [-81.410281756962192, 31.299458077094936],
-              [-81.409006186321605, 31.300902119329557],
-              [-81.406069967111222, 31.302851576346292],
-              [-81.404649992247172, 31.30424748383976],
-              [-81.402416399783391, 31.307194055564054],
-              [-81.408068869526204, 31.311080128512241],
-              [-81.414239629105893, 31.306596335656081],
-              [-81.42004558, 31.31077189],
-              [-81.435504726373594, 31.314279622117649],
-              [-81.436341483157221, 31.316026428249913],
-              [-81.436335, 31.317257],
-              [-81.43436, 31.329013],
-              [-81.43484, 31.330984],
-              [-81.435374, 31.331649],
-              [-81.437083, 31.332749],
-              [-81.439219, 31.33323],
-              [-81.442449, 31.332864],
-              [-81.445718868903541, 31.329777639133908],
-              [-81.452607870287949, 31.327054843348641],
-              [-81.454881274920893, 31.327298834446459],
-              [-81.461847644335634, 31.333974471863183],
-              [-81.463377422734666, 31.336012571426178],
-              [-81.46616, 31.336985],
-              [-81.466747, 31.337512],
-              [-81.466908, 31.338337],
-              [-81.467469, 31.339185],
-              [-81.468136, 31.339804],
-              [-81.470198633376853, 31.340351670379157],
-              [-81.471047, 31.340559],
-              [-81.472382, 31.339734],
-              [-81.472782, 31.339253],
-              [-81.473075, 31.338405],
-              [-81.473048, 31.337327],
-              [-81.471026003939912, 31.334486532387633],
-              [-81.471026003939912, 31.333567231762],
-              [-81.472276252790778, 31.333089195436674],
-              [-81.473820677841829, 31.333990110049793],
-              [-81.474035, 31.335494],
-              [-81.473983, 31.338083],
-              [-81.474277, 31.340031],
-              [-81.474518, 31.340077],
-              [-81.475209, 31.339634],
-              [-81.47638728, 31.33941482],
-              [-81.47665405, 31.33964157],
-              [-81.47801971, 31.33867836],
-              [-81.47826385, 31.33795929],
-              [-81.4784966, 31.33781216],
-              [-81.47926976, 31.33768483],
-              [-81.479885, 31.337944],
-              [-81.48028564, 31.33789825],
-              [-81.48065948, 31.33745194],
-              [-81.48226882, 31.33718346],
-              [-81.48260498, 31.33748436],
-              [-81.48255157, 31.33787346],
-              [-81.48140717, 31.33966255],
-              [-81.4815979, 31.34000588],
-              [-81.482182, 31.340212],
-              [-81.48431428, 31.33984592],
-              [-81.48447418, 31.33998108],
-              [-81.484479, 31.341517],
-              [-81.48471832, 31.34220314],
-              [-81.48514557, 31.34257126],
-              [-81.486082, 31.34289101],
-              [-81.48870087, 31.34206581],
-              [-81.48918152, 31.3420887],
-              [-81.489153, 31.343692],
-              [-81.48934, 31.344654],
-              [-81.49005788, 31.34423906],
-              [-81.49012999, 31.343456],
-              [-81.4929966, 31.34501216],
-              [-81.4966966, 31.34741216],
-              [-81.5026966, 31.35251216],
-              [-81.5056966, 31.35421216],
-              [-81.5089966, 31.35491216],
-              [-81.5175966, 31.35541216],
-              [-81.5191966, 31.35641216],
-              [-81.5209966, 31.35721216],
-              [-81.5230966, 31.36031216],
-              [-81.5254966, 31.36341216],
-              [-81.5279966, 31.36511216],
-              [-81.5306966, 31.36621216],
-              [-81.5358966, 31.36761216],
-              [-81.5375966, 31.37031216],
-              [-81.5386966, 31.37301216],
-              [-81.5394966, 31.37611216],
-              [-81.5422966, 31.37691216],
-              [-81.5462966, 31.37631216],
-              [-81.54889144, 31.375224],
-              [-81.548598, 31.375224],
-              [-81.548868, 31.374966],
-              [-81.549448, 31.374834],
-              [-81.549925, 31.374934],
-              [-81.550307, 31.375224],
-              [-81.54987324, 31.375224],
-              [-81.5520966, 31.37621216],
-              [-81.5520966, 31.37891216],
-              [-81.5486966, 31.38051216],
-              [-81.5469966, 31.38371216],
-              [-81.5512966, 31.38591216],
-              [-81.5584966, 31.38611216],
-              [-81.5630966, 31.38891216],
-              [-81.5657966, 31.39111216],
-              [-81.5703966, 31.39321216],
-              [-81.5730966, 31.39501216],
-              [-81.5776966, 31.39871216],
-              [-81.5803966, 31.40141216],
-              [-81.5810966, 31.40511216],
-              [-81.5829966, 31.40741216],
-              [-81.5852966, 31.40901216],
-              [-81.5887966, 31.40991216],
-              [-81.5905966, 31.41321216],
-              [-81.5927966, 31.41491216],
-              [-81.5951966, 31.41501216],
-              [-81.5961966, 31.41771216],
-              [-81.5988966, 31.42131216],
-              [-81.6034966, 31.42381216],
-              [-81.6059966, 31.42631216],
-              [-81.6062966, 31.42831216],
-              [-81.6059966, 31.43051216],
-              [-81.6076966, 31.43221216],
-              [-81.6107966, 31.43391216],
-              [-81.6120966, 31.43571216],
-              [-81.61049659, 31.43811216],
-              [-81.6077966, 31.43931216],
-              [-81.6099966, 31.44141216],
-              [-81.6112966, 31.44321216],
-              [-81.6121966, 31.44561216],
-              [-81.6162966, 31.44571216],
-              [-81.6188966, 31.44631216],
-              [-81.6207966, 31.44741216],
-              [-81.6249966, 31.45001216],
-              [-81.625, 31.45311165],
-              [-81.62049866, 31.45591164],
-              [-81.6171966, 31.45851216],
-              [-81.6160965, 31.46441269],
-              [-81.6133966, 31.46751216],
-              [-81.6125966, 31.46991216],
-              [-81.61759949, 31.47851181],
-              [-81.6189966, 31.48041216],
-              [-81.6207966, 31.48241216],
-              [-81.62339783, 31.48461151],
-              [-81.62669373, 31.48411179],
-              [-81.63209534, 31.48371124],
-              [-81.6395966, 31.48761216],
-              [-81.6422966, 31.48951216],
-              [-81.6473999, 31.4956131],
-              [-81.646698, 31.49771118],
-              [-81.64550018, 31.49931145],
-              [-81.6459966, 31.50151216],
-              [-81.6480966, 31.50251216],
-              [-81.6515966, 31.50341216],
-              [-81.6545966, 31.50561216],
-              [-81.65679932, 31.5078125],
-              [-81.6582966, 31.51061216],
-              [-81.65769958, 31.51411247],
-              [-81.6552966, 31.51841216],
-              [-81.65570068, 31.52271271],
-              [-81.65499878, 31.52531242],
-              [-81.6526966, 31.52811216],
-              [-81.6547966, 31.52941216],
-              [-81.6569966, 31.52851216],
-              [-81.6598966, 31.52791216],
-              [-81.6630966, 31.52901216],
-              [-81.66729736, 31.53261185],
-              [-81.66809845, 31.53601265],
-              [-81.66479492, 31.53831291],
-              [-81.66179657, 31.53911209],
-              [-81.62449646, 31.55091286],
-              [-81.56659698, 31.5771122],
-              [-81.54429626, 31.61421204],
-              [-81.5331955, 31.6321125],
-              [-81.51439667, 31.66411209],
-              [-81.49186805, 31.69941952],
-              [-81.48633, 31.699515],
-            ],
-          ],
-        ],
-      },
-    },
-  ],
-};
-
-export const counties: StyleSpecification = {
-  version: 8,
-  name: "County",
-  sources: {
-    counties: {
-      type: "geojson",
-      data: simpleCounties,
-      promoteId: "uuid",
-    },
-  },
-  layers: [
-    {
-      id: "counties-fill",
-      type: "fill",
-      source: "counties",
-      layout: {
-        visibility: "none",
-      },
-      paint: {
-        "fill-color": "#ea580c",
-        "fill-opacity": [
-          "case",
-          ["boolean", ["feature-state", "hovered"], false],
-          0.65,
-          0.25,
-        ],
-      },
-    },
-    {
-      id: "counties-outline",
-      type: "line",
-      source: "counties",
-      layout: {
-        "line-join": "round",
-        "line-cap": "round",
-        visibility: "none",
-      },
-      paint: {
-        "line-color": "#ea580c",
-        "line-width": 1,
-        "line-opacity": 0.25,
-      },
-    },
-  ],
-};
