+++ app/mapStyles/index.ts
@@ -2,9 +2,7 @@ export * from "./base";
 export * from "./satellite";
 export * from "./labels";
 export * from "./usgs";
-export * from "./combined";
-export * from "./counties";
-export * from "./islands";
+export * from "./mapStyle";
 export * from "./masks";
 export * from "./costalLabels";
 export * from "./wms";
