--- app/mapStyles/sources/islands.ts
@@ -1,1043 +1,0 @@
-import type { FeatureCollection } from "geojson";
-
-export const simpleIslandShapes: FeatureCollection = {
-  type: "FeatureCollection",
-  features: [
-    {
-      type: "Feature",
-      properties: {
-        name: "Blackbeard Island",
-        description:
-          '<p>Blackbeard Island is northeast of Sapelo Island. It is named after Edward Teach (d. 1718), nicknamed "Blackbeard," an early pirate who used this land. Edward Teach established headquarters on the island in 1716. Now, it is federally owned and contains the Blackbeard Island National Wildlife Refuge.</p>',
-        uuid: "c4e43063-179b-5807-a191-6e1438170486",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-81.203524440061088, 31.467500507733988],
-            [-81.229582276220611, 31.443041392209967],
-            [-81.231627043113264, 31.442267696628964],
-            [-81.233948129856273, 31.441991376778606],
-            [-81.235992896748925, 31.443925615731114],
-            [-81.234003393826342, 31.448015149516419],
-            [-81.231019139442466, 31.452933642852795],
-            [-81.227813829178316, 31.453983658284159],
-            [-81.223724295393012, 31.454646825925018],
-            [-81.22167952850036, 31.45575210532645],
-            [-81.220406169433076, 31.458200399733123],
-            [-81.221305846521332, 31.462364042072508],
-            [-81.225067720876424, 31.466125916427597],
-            [-81.228761961306645, 31.469195405672181],
-            [-81.22800095380552, 31.470548307896408],
-            [-81.223716763428797, 31.474776127347123],
-            [-81.221546482777427, 31.475452578459237],
-            [-81.219383985890047, 31.477662658856616],
-            [-81.218866938305297, 31.479010675774013],
-            [-81.218738227812935, 31.480621431036262],
-            [-81.218341078317906, 31.482487475720056],
-            [-81.21869094508736, 31.483316789543945],
-            [-81.219079685942305, 31.484521886194287],
-            [-81.218120791833442, 31.484819920849748],
-            [-81.215166361335818, 31.484418221966301],
-            [-81.212730251978144, 31.483640740256405],
-            [-81.209322290483087, 31.483148335173468],
-            [-81.207080551552892, 31.483537076028416],
-            [-81.206264195757498, 31.484794004792754],
-            [-81.207493235280381, 31.487047741219012],
-            [-81.218085631647355, 31.494899393408339],
-            [-81.219761003810135, 31.495062006116083],
-            [-81.220444170346568, 31.495296595011812],
-            [-81.221712028471188, 31.495727798844456],
-            [-81.22244681762227, 31.495977255601606],
-            [-81.222864642623421, 31.496116533322748],
-            [-81.223268332661576, 31.496445473626608],
-            [-81.223932076208328, 31.497758928355221],
-            [-81.224549641935866, 31.498254177831363],
-            [-81.225392387782634, 31.498527595966845],
-            [-81.226293772264484, 31.498812883760753],
-            [-81.226779622944719, 31.498968697698182],
-            [-81.227075751620731, 31.499132697144855],
-            [-81.22716416219437, 31.49941577801037],
-            [-81.227020534469276, 31.500118097262952],
-            [-81.226420553544557, 31.500843434678519],
-            [-81.226097755217154, 31.501217850876735],
-            [-81.225024537728061, 31.502245155008627],
-            [-81.224612133752288, 31.502451250658901],
-            [-81.224298589361439, 31.502442245541509],
-            [-81.222674462135188, 31.500671269972099],
-            [-81.222351109004535, 31.500746670475014],
-            [-81.2221337860272, 31.501295564434905],
-            [-81.222154258280852, 31.503077583509583],
-            [-81.221793589728705, 31.503576827059042],
-            [-81.220443638984918, 31.503842123947528],
-            [-81.219914400373327, 31.504100488589266],
-            [-81.219052967682842, 31.505300519932984],
-            [-81.215621902126188, 31.506845063922512],
-            [-81.215198839556265, 31.50721524367119],
-            [-81.215077816608641, 31.50781968542443],
-            [-81.214797381548934, 31.508526037278379],
-            [-81.214461024795014, 31.509391514664706],
-            [-81.213410006689571, 31.510323787784507],
-            [-81.212514571786713, 31.511121250235064],
-            [-81.212177156527687, 31.511735491106908],
-            [-81.212291884592901, 31.512193785940514],
-            [-81.213202931376358, 31.512547593767913],
-            [-81.214663410668564, 31.512706280847553],
-            [-81.215187784270995, 31.512939675559348],
-            [-81.215689601487171, 31.513168660114303],
-            [-81.216270356950858, 31.513784998518471],
-            [-81.216451964416052, 31.514097136349275],
-            [-81.216764102246856, 31.514681685377866],
-            [-81.217033285289219, 31.515376795451939],
-            [-81.21811503558196, 31.518149802630344],
-            [-81.218529870933253, 31.519087505278932],
-            [-81.219719760439403, 31.521569405918008],
-            [-81.219740252014233, 31.5221154827315],
-            [-81.218531719727366, 31.522781975794704],
-            [-81.215663706706678, 31.525084465121175],
-            [-81.213366626783767, 31.528790531272858],
-            [-81.204117931462918, 31.537516558046224],
-            [-81.198406477898573, 31.538587455589536],
-            [-81.187162053693783, 31.533411450796855],
-            [-81.173169451380602, 31.517828444702072],
-            [-81.203524440061088, 31.467500507733988],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Cumberland Island",
-        description:
-          "<p>Located off Georgia's southeastern coast and accessible only by boat, Cumberland Island is home to wildlife, including wild horses, loggerhead sea turtles, and a range of bird species. The island’s history spans Indigenous Timucua settlements, Spanish missions, and English colonial outposts, with significant remnants from the 19th century, when the Carnegie family built opulent estates like the now-ruined Dungeness mansion.</p>",
-        uuid: "ab536bff-55c5-50d5-b06f-2be190dd88a4",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-81.400238909679871, 30.899971443604628],
-            [-81.444920337503532, 30.705909590139942],
-            [-81.467732003121256, 30.715886564686521],
-            [-81.485333432326613, 30.727408085284111],
-            [-81.482248335959383, 30.741324709850474],
-            [-81.470422028934763, 30.771697412574557],
-            [-81.471036095803726, 30.779680281870995],
-            [-81.480758821228861, 30.791040518946691],
-            [-81.480217964405213, 30.795773507109207],
-            [-81.476664208498264, 30.801741134952941],
-            [-81.474362509518627, 30.811036781934234],
-            [-81.472650583496048, 30.814762738571616],
-            [-81.474110755691768, 30.818941852097325],
-            [-81.480807407485983, 30.822164301081006],
-            [-81.484618464766882, 30.83257043279685],
-            [-81.47983222005773, 30.845493293511549],
-            [-81.475440679543354, 30.850004555605594],
-            [-81.472065058946299, 30.851423364031312],
-            [-81.469702932461473, 30.852233100522025],
-            [-81.468056509487297, 30.85395905447805],
-            [-81.466685163619744, 30.857339870791758],
-            [-81.467282929767137, 30.86044890916126],
-            [-81.469181716352992, 30.86280325949291],
-            [-81.47236887948732, 30.864616507622326],
-            [-81.475867072201552, 30.865348157376165],
-            [-81.477785353724215, 30.866945634598302],
-            [-81.478775250471031, 30.869791061801504],
-            [-81.481261502765278, 30.871095187707091],
-            [-81.483345627859293, 30.870687448792125],
-            [-81.487577485554183, 30.865766114813923],
-            [-81.492150622095437, 30.866439883798193],
-            [-81.494744341327788, 30.867640938503406],
-            [-81.497269804790861, 30.872005620345824],
-            [-81.497596814210794, 30.875012528683857],
-            [-81.490793559523752, 30.880442791018666],
-            [-81.482024126116912, 30.882806304808827],
-            [-81.468070688869034, 30.885580820380071],
-            [-81.449627529633162, 30.887519318303166],
-            [-81.446893742213092, 30.889991504433404],
-            [-81.445467265714782, 30.896111995677511],
-            [-81.462584983694654, 30.90671992040717],
-            [-81.469975208832906, 30.915217863664846],
-            [-81.455128193008534, 30.92645344320761],
-            [-81.436154997313636, 30.945425655773619],
-            [-81.42688192617122, 30.978786227623775],
-            [-81.419519275512826, 30.98104997454201],
-            [-81.397203349259399, 30.977095253433806],
-            [-81.400238909679871, 30.899971443604628],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Tybee Island",
-        description:
-          "<p>Tybee belongs to the ancestral land of the Yuchi (Euchee) tribe now based in Oklahoma. Following occupation by the Spanish and French, it became a military outpost, its positioning along the Savannah River making it a strategic location during the Revolutionary War, Spanish-American War, Civil War, and World Wars I and II. Within the past two centuries it has developed into a popular tourist destination and features one of the oldest operating lighthouses in the country.</p>",
-        uuid: "532b46db-0bf7-486b-b055-3b852ee6d21f",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-80.851002335587765, 32.030258401293324],
-            [-80.855592853844598, 32.028389474560882],
-            [-80.86082140259802, 32.023204497047075],
-            [-80.876724905056321, 32.017540235897542],
-            [-80.882196545280408, 32.017327621859963],
-            [-80.883532193072412, 32.016035527800312],
-            [-80.883662854269446, 32.014119163577007],
-            [-80.882244588404404, 32.011593548263406],
-            [-80.881607314353374, 32.011726710005412],
-            [-80.882001762239881, 32.013749868850873],
-            [-80.881085298494853, 32.013817470811823],
-            [-80.880493338572961, 32.013922419005283],
-            [-80.878729850916187, 32.014742755568555],
-            [-80.878449426634916, 32.014818614059813],
-            [-80.878453070035405, 32.015315675254719],
-            [-80.87843747787629, 32.015533089769576],
-            [-80.87226835569966, 32.015235185193212],
-            [-80.871339243395184, 32.015221450421897],
-            [-80.870165977991419, 32.015498101866584],
-            [-80.866244604320613, 32.017169463689974],
-            [-80.865520101436545, 32.016541512252246],
-            [-80.86390298887099, 32.015484090056852],
-            [-80.863857420030186, 32.015372372898739],
-            [-80.864226380644453, 32.015062210788734],
-            [-80.86417961604738, 32.014957891436232],
-            [-80.863991877219931, 32.014841672162092],
-            [-80.863902477778296, 32.014507914246622],
-            [-80.863751490709944, 32.014490380448031],
-            [-80.863171028129557, 32.014469822398311],
-            [-80.862958633332241, 32.01438953564935],
-            [-80.862741891342225, 32.014495545574839],
-            [-80.862381745032508, 32.0144068179928],
-            [-80.862183991563526, 32.01442566958319],
-            [-80.861970301054683, 32.014502697557305],
-            [-80.861581832693076, 32.014320057389135],
-            [-80.861453732413651, 32.012492170078701],
-            [-80.861857893276778, 32.011453879541911],
-            [-80.862563030554043, 32.011084251130441],
-            [-80.864411172611398, 32.010896593629226],
-            [-80.864257634655857, 32.010731682491802],
-            [-80.863086196921046, 32.010856787492607],
-            [-80.862153596005953, 32.010851100901661],
-            [-80.860321791427253, 32.010002849315669],
-            [-80.859720764963967, 32.009389599508836],
-            [-80.859376835205509, 32.00832710221934],
-            [-80.861655651854861, 32.005960003390221],
-            [-80.860326081417426, 32.004584005577421],
-            [-80.859465923033824, 32.002490226751185],
-            [-80.860182183036244, 32.000739785430902],
-            [-80.856794678481975, 31.99887246843279],
-            [-80.855515756113689, 31.997575275744978],
-            [-80.853590139540017, 31.984381456002549],
-            [-80.847309276459683, 31.98550677730444],
-            [-80.837803073876586, 32.001854364427793],
-            [-80.839780949985112, 32.011429296638035],
-            [-80.840827011354861, 32.020463463013051],
-            [-80.840541721890375, 32.022270296288049],
-            [-80.841017204331166, 32.024552612003845],
-            [-80.844821063857495, 32.025408480397267],
-            [-80.846247511179854, 32.025693769861739],
-            [-80.851002335587765, 32.030258401293324],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Ossabaw Island",
-        description:
-          "<p>An ancestral home of several Native American tribes, including the Creek and the Guale, Ossabaw Island has supported migratory and permanent settlements for 4,000 years. Violence and pressure from European settlers forced indigenous residents to relocate further inland, but their material legacy is captured - partially - in the many archaeological sites on Ossabaw. By the 19th century the island was developed as a cotton and indigo plantation. Sweetfield Church in Pin Point, GA, originated from the historic Hinder Me Not Baptist Church established on Ossabaw by the African American community residing there in 1878.</p>",
-        uuid: "79b2c613-a9f7-4e1e-b795-0faeed53c827",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-81.037445206804364, 31.805481728393254],
-            [-81.096348142316131, 31.743579563883852],
-            [-81.120758608844724, 31.725388529309921],
-            [-81.14009498624398, 31.719163805078654],
-            [-81.151352466236702, 31.745387111414626],
-            [-81.17043814955538, 31.755503469101729],
-            [-81.171878830054979, 31.757639650532177],
-            [-81.168401325400765, 31.767326984926065],
-            [-81.157888411159092, 31.78422827780323],
-            [-81.161943859817413, 31.789983192417097],
-            [-81.16620019325579, 31.791840350499751],
-            [-81.177580289858909, 31.792194779386605],
-            [-81.180653262158401, 31.794252825055068],
-            [-81.178933525366944, 31.798453493611248],
-            [-81.174960744040135, 31.806924688282265],
-            [-81.175191317099504, 31.816245665546131],
-            [-81.174240957175925, 31.817736747495211],
-            [-81.158690534255683, 31.822630777687611],
-            [-81.155944516564148, 31.837417026795883],
-            [-81.140718340716973, 31.849381998565047],
-            [-81.131336150907586, 31.852722973472776],
-            [-81.125186638493574, 31.850844839015011],
-            [-81.112684246892869, 31.851411549718929],
-            [-81.093076865103114, 31.846940281090205],
-            [-81.085152450770536, 31.837665810073982],
-            [-81.079311590942069, 31.831637145029841],
-            [-81.067503519766902, 31.82775233191602],
-            [-81.053848945966678, 31.820925045015912],
-            [-81.044897614253202, 31.826993744482674],
-            [-81.03321536777969, 31.821531914962588],
-            [-81.037445206804364, 31.805481728393254],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Sea Island",
-        description:
-          "<p>Sea Island is a private-owned unincorporated place, and it is part of the Golden Isles of Georgia.</p>",
-        uuid: "3a1c0b63-b2b4-45df-a5e1-33cc8a0c5469",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-81.307593454750503, 31.211227928403677],
-            [-81.359225525688984, 31.159803765440198],
-            [-81.360688456195533, 31.164070236372552],
-            [-81.358651115486651, 31.168683444408934],
-            [-81.355944219903904, 31.172275286624508],
-            [-81.353081157268292, 31.176335629998643],
-            [-81.351519486739775, 31.179250748318537],
-            [-81.351519486739775, 31.180916530215622],
-            [-81.352977045899721, 31.182269978007],
-            [-81.353377454120064, 31.183352110099463],
-            [-81.352609733741957, 31.184294924598884],
-            [-81.350670229628875, 31.185264676655432],
-            [-81.348147709564643, 31.186167886924192],
-            [-81.349099834491128, 31.188979251618935],
-            [-81.348660939265898, 31.189724214068626],
-            [-81.345981406895987, 31.193377713832476],
-            [-81.346699347023659, 31.195707531856041],
-            [-81.348901200821999, 31.19609609429104],
-            [-81.350300728172968, 31.196661014416595],
-            [-81.351852003281806, 31.199056962883116],
-            [-81.352587147736116, 31.200007530558338],
-            [-81.351150770082555, 31.201101913532472],
-            [-81.350124786044304, 31.202452792516173],
-            [-81.34957776106954, 31.203784760267027],
-            [-81.348534115873633, 31.206738124375892],
-            [-81.349418407183222, 31.210275289614259],
-            [-81.350155316607882, 31.217673860237841],
-            [-81.350906749631932, 31.221482496178108],
-            [-81.349898232823776, 31.221752215789593],
-            [-81.340259861759364, 31.219704051188454],
-            [-81.336658004635495, 31.217720947778265],
-            [-81.330901432348995, 31.21549259721575],
-            [-81.326537579164068, 31.21688531631732],
-            [-81.325237708002604, 31.222084800963191],
-            [-81.326073339463548, 31.224684543286124],
-            [-81.328208842085957, 31.231183899093462],
-            [-81.324959164182289, 31.233133705835662],
-            [-81.318181264554639, 31.224127455645498],
-            [-81.307593454750503, 31.211227928403677],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Little Tybee Island",
-        description:
-          "<p>Little Tybee Island is owned by the State of Georgia and covers approximately 6,780 acres, of which only 600 acres are upland. Despite its name, it is twice the area of Tybee Island to the north and includes salt marshes and sand dunes. It receives relatively few visitors due to a lack of road access and permanent structures, though camping and recreation are permitted for those who reach the island by boat. Many bird species nest in Little Tybee's marshes and forests.</p>",
-        uuid: "e2cc0661-11c1-47d3-88a4-708721a59f3e",
-      },
-      geometry: {
-        type: "MultiPolygon",
-        coordinates: [
-          [
-            [
-              [-80.85643482963998, 31.994814861403533],
-              [-80.858688, 31.980002],
-              [-80.859794582999939, 31.976267785000061],
-              [-80.860107421999942, 31.975097656000059],
-              [-80.860900878999985, 31.973327637000068],
-              [-80.861328124999943, 31.972900391000053],
-              [-80.861877441, 31.971679688000052],
-              [-80.862609862999989, 31.969482422000056],
-              [-80.865905761999954, 31.967895508000026],
-              [-80.890928079672108, 31.946036629592392],
-              [-80.911509898481611, 31.938013208700557],
-              [-80.925177885338925, 31.940420805322116],
-              [-80.928952314503462, 31.942208692821108],
-              [-80.932312011999954, 31.961730957000043],
-              [-80.931213378999985, 31.962707520000038],
-              [-80.929809569999975, 31.963500977000024],
-              [-80.927917479999962, 31.965515137000068],
-              [-80.926086425999983, 31.966674805000025],
-              [-80.925598144999981, 31.967285156000059],
-              [-80.925109862999989, 31.968078613000046],
-              [-80.924377441, 31.970092773000033],
-              [-80.924013757773324, 31.972062807883894],
-              [-80.922485351999967, 31.974670410000044],
-              [-80.922607421999942, 31.976928711000028],
-              [-80.923532134794826, 31.979168198624606],
-              [-80.925109862999989, 31.982482910000044],
-              [-80.926208495999958, 31.983703613000046],
-              [-80.927356643230979, 31.985291591913043],
-              [-80.927185058999953, 31.990722656000059],
-              [-80.927033652280258, 31.994233872366951],
-              [-80.925720214999956, 31.993896484000061],
-              [-80.923217772999976, 31.992675781000059],
-              [-80.920410155999946, 31.990722656000059],
-              [-80.919311522999976, 31.989501953000058],
-              [-80.917494452556298, 31.98781754138275],
-              [-80.914535608766812, 31.988430231258352],
-              [-80.913879394999981, 31.989074707000043],
-              [-80.91320800799997, 31.991699219000054],
-              [-80.912109374999943, 31.992919922000056],
-              [-80.909484862999989, 31.994873047000056],
-              [-80.909118651999961, 31.994506836000028],
-              [-80.906677245999958, 31.993713379000042],
-              [-80.905190974558977, 31.993887886320628],
-              [-80.903381347999982, 31.994323730000076],
-              [-80.900915034864482, 31.996224224504218],
-              [-80.898446451123334, 31.995827487831534],
-              [-80.892054582507839, 31.993050331122738],
-              [-80.890502929999968, 31.994506836000028],
-              [-80.889709472999982, 31.994873047000056],
-              [-80.888183593999941, 31.996093750000057],
-              [-80.886108397999976, 31.997314453000058],
-              [-80.884826659999987, 31.997680664000029],
-              [-80.883605956999986, 31.997680664000029],
-              [-80.882385253999985, 31.996093750000057],
-              [-80.880981444999975, 31.993530273000033],
-              [-80.880187987999989, 31.992675781000059],
-              [-80.878784179999968, 31.991882324000073],
-              [-80.877502441, 31.991516113000046],
-              [-80.875610351999967, 31.991516113000046],
-              [-80.874023436999948, 31.991882324000073],
-              [-80.872375487999989, 31.992675781000059],
-              [-80.868307830616786, 31.995753153274041],
-              [-80.861511229999962, 31.998474121000072],
-              [-80.859321881549235, 31.999037174820817],
-              [-80.85643482963998, 31.994814861403533],
-            ],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Wolf Island",
-        description:
-          "<p>One of three islands contained by the Wolf Island National Wildlife Refuge. Though its waterways are open for recreation, most of the island is closed to the public to provide habitats for local wildlife, including loggerhead turtles, ducks, and a variety of seabirds.</p>",
-        uuid: "dcec5413-40a6-418a-b5f5-db04e183cf96",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-81.32687733363187, 31.365374450750313],
-            [-81.322954874920001, 31.369215110742832],
-            [-81.318531211965961, 31.369447935108838],
-            [-81.311013183910546, 31.366790333838026],
-            [-81.289861010605733, 31.367502688034026],
-            [-81.284455394492028, 31.348156272469215],
-            [-81.279618790600821, 31.326249301903179],
-            [-81.292137059495701, 31.32112819190073],
-            [-81.305182614981845, 31.322184588810408],
-            [-81.321701307849736, 31.329390631083871],
-            [-81.326745835779775, 31.333814294037911],
-            [-81.325504105827775, 31.341807930603984],
-            [-81.328065173853787, 31.349723959048056],
-            [-81.336655957003956, 31.359088192868256],
-            [-81.335802267661947, 31.362658166480287],
-            [-81.32687733363187, 31.365374450750313],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Jekyll Island",
-        description:
-          "<p>Jekyll Island was originally named Ospo by the Native Americans, Gualdaquini by the Spanish, and Isle de la Somme by the French. In January 1793, General James E. Oglethorpe named the island after Sir Joesph Jekyll (1663-1738) who was a lawyer and statesman who helped finance the colonial venture.</p><p><br></p><p>The Spanish first came on the island around 1556 and maintained a Jesuit mission until around 1742. In 1791, four Frenchmen purchased the island, and the du Bignon family owned Jekyll Island until 1886. The first Englishman to settle here was Major William Horton; he established a brewery to supply the Frederica community on St. Simons Island nearby.</p><p><br></p><p>From 1888 to 1942, millionaires including Astor, Morgan, Pulitzer, Rockefeller, and Vanderbilt maintained cottages here. In 1947, the State of Georgia bought the island for $675,000 to establish the Jekyll Island State Park with 11,000 acres consisting of ten miles of white sandy beach.</p>",
-        uuid: "aa8a4b3b-febb-52fd-acaf-7e5f73ba1426",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-81.440715892634557, 31.090690619833296],
-            [-81.424943030054052, 31.105972405548702],
-            [-81.416814671164488, 31.122985249736146],
-            [-81.404527617029117, 31.12109493371532],
-            [-81.39992043836294, 31.093904181142214],
-            [-81.401829350806466, 31.057185688846033],
-            [-81.421633287328206, 31.008310622846917],
-            [-81.438239608803983, 31.00743660592714],
-            [-81.435836062274589, 31.031690575450973],
-            [-81.425531348467771, 31.038530492854264],
-            [-81.425814959734325, 31.06184407989851],
-            [-81.428818205964234, 31.068080451483983],
-            [-81.430848813886129, 31.07655433454266],
-            [-81.432909651788037, 31.083263570561151],
-            [-81.434979309862285, 31.086270432291649],
-            [-81.437358053482527, 31.08817224046927],
-            [-81.439720977330254, 31.089478066806173],
-            [-81.440715892634557, 31.090690619833296],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Saint Simons Island",
-        description:
-          "<p>Saint Simons Island is famous for beaches, which is also a great place for relaxation and adventure.</p>",
-        uuid: "8c113813-f475-42b9-bcfa-4915d24009a9",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-81.316191974750424, 31.27158031149397],
-            [-81.313978935056312, 31.266938983077399],
-            [-81.314785461332306, 31.265783688682053],
-            [-81.318491122600392, 31.265543910600002],
-            [-81.323308482248905, 31.266023466764107],
-            [-81.330224614980622, 31.26292619509838],
-            [-81.334123789928711, 31.252381094552756],
-            [-81.335441256537266, 31.24801698641194],
-            [-81.334123789928711, 31.244229269912363],
-            [-81.326960065244734, 31.239865161771547],
-            [-81.325560256973148, 31.235089345315558],
-            [-81.327623850188814, 31.233834972576801],
-            [-81.329651669560548, 31.231882257626236],
-            [-81.330387692888067, 31.229433853495909],
-            [-81.32878045827492, 31.226369593111947],
-            [-81.32626128357326, 31.219499436525066],
-            [-81.327717927082659, 31.217651455953437],
-            [-81.330031989713959, 31.216620382356844],
-            [-81.334233980132453, 31.218586120363778],
-            [-81.34218710362839, 31.221525710227354],
-            [-81.350953281166454, 31.222091382210618],
-            [-81.352757753207598, 31.220286910169481],
-            [-81.350394754106105, 31.211522331683963],
-            [-81.349506836314177, 31.20700341176877],
-            [-81.35039112762378, 31.203348341022458],
-            [-81.351468844606785, 31.201168099257441],
-            [-81.35344517112803, 31.200265863236872],
-            [-81.350444075957228, 31.196418744159079],
-            [-81.348754039202475, 31.195479834850886],
-            [-81.347026446075404, 31.195235718430755],
-            [-81.346200205884202, 31.193620794420667],
-            [-81.348114301079931, 31.190517761808501],
-            [-81.349241031894692, 31.189288600919674],
-            [-81.349343461968758, 31.187683863092591],
-            [-81.348555576802937, 31.186644467225047],
-            [-81.348826759271603, 31.186322832204063],
-            [-81.350006087681876, 31.185648029316898],
-            [-81.350920540192519, 31.185301168019759],
-            [-81.352726167027427, 31.184606340421077],
-            [-81.353562452478997, 31.183256546358884],
-            [-81.353811965855058, 31.182100041800222],
-            [-81.35317500785689, 31.181359090659498],
-            [-81.352330063573604, 31.180917119803624],
-            [-81.351719103861072, 31.179929184949323],
-            [-81.351800492758713, 31.179492109395888],
-            [-81.351923163352467, 31.178989631771508],
-            [-81.356063789793794, 31.173410384612957],
-            [-81.358196835709592, 31.169997511147674],
-            [-81.359618866320133, 31.167615609875028],
-            [-81.360936390356812, 31.164773230705535],
-            [-81.361863140741562, 31.163904402219824],
-            [-81.362650138446526, 31.162620916397941],
-            [-81.363547936340538, 31.161565610101459],
-            [-81.364749194608805, 31.160676653629046],
-            [-81.365770509023307, 31.159928428640878],
-            [-81.366090493904963, 31.159298458405111],
-            [-81.365272940084921, 31.156003232457575],
-            [-81.36531053075592, 31.155078657778226],
-            [-81.363047400979013, 31.153833184387523],
-            [-81.359313835177119, 31.155022638979272],
-            [-81.360051737562742, 31.152654743264211],
-            [-81.366220905301475, 31.143045560793229],
-            [-81.377151706337756, 31.133219397244225],
-            [-81.402253480727026, 31.129539555972499],
-            [-81.418024229034415, 31.137424930126198],
-            [-81.418905966917265, 31.153401293957817],
-            [-81.417481000893886, 31.160639894046565],
-            [-81.414389913812983, 31.165007255614348],
-            [-81.408647486045197, 31.172532874174131],
-            [-81.413267715156934, 31.182287958772562],
-            [-81.418431068569006, 31.184246472135765],
-            [-81.418524526951899, 31.189772856778454],
-            [-81.413759046545607, 31.197498711376536],
-            [-81.412387165822594, 31.20262521302573],
-            [-81.410076629868016, 31.203275051262949],
-            [-81.404228085733024, 31.202119783285667],
-            [-81.396791048129259, 31.203563868257273],
-            [-81.395635780151977, 31.207751714674924],
-            [-81.401989754027042, 31.214683322538622],
-            [-81.401484324286983, 31.216416224504545],
-            [-81.397422719566649, 31.21613146983022],
-            [-81.394972140043933, 31.218482701534448],
-            [-81.393746850282582, 31.223416976519374],
-            [-81.394935204115981, 31.224170191146122],
-            [-81.399136075178234, 31.224435096662219],
-            [-81.397652214149915, 31.232213863222071],
-            [-81.397902061628656, 31.235162063471286],
-            [-81.399251238013889, 31.238160233216252],
-            [-81.398401756586154, 31.240308921533476],
-            [-81.396153129277423, 31.241658097918709],
-            [-81.393330090877228, 31.239729114875644],
-            [-81.392268524286521, 31.238945577630119],
-            [-81.390398145055272, 31.239324708555372],
-            [-81.389382517315838, 31.242721265877091],
-            [-81.38795506914056, 31.243336006332132],
-            [-81.384438263930093, 31.245294867445974],
-            [-81.384807626329163, 31.2481938026386],
-            [-81.385882135126423, 31.249380239435581],
-            [-81.386803224741143, 31.250814522461788],
-            [-81.386883653393113, 31.25435338314864],
-            [-81.387889011542782, 31.25564024158022],
-            [-81.393418481365984, 31.259179102267073],
-            [-81.393860838951852, 31.261612068989283],
-            [-81.389547555499121, 31.266026743604527],
-            [-81.383572610800599, 31.274010073075676],
-            [-81.382919885749487, 31.280939000541199],
-            [-81.380638395352193, 31.283670939126086],
-            [-81.378658264701428, 31.289260017575849],
-            [-81.380350957031922, 31.293411904424243],
-            [-81.377268667672837, 31.295196683006445],
-            [-81.373691988378141, 31.299942661301333],
-            [-81.370570518061356, 31.295647028020333],
-            [-81.364337596376615, 31.294852771535801],
-            [-81.358491101375677, 31.297672845595081],
-            [-81.356083721081163, 31.299358011801239],
-            [-81.349754767227452, 31.298011123047676],
-            [-81.345946911442667, 31.297271409948831],
-            [-81.342503346365987, 31.296197594014757],
-            [-81.334702441166002, 31.291626095806166],
-            [-81.326440192110098, 31.289780720382513],
-            [-81.316116362842493, 31.282374341246427],
-            [-81.316942784303578, 31.276431612546379],
-            [-81.316191974750424, 31.27158031149397],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Little Saint Simons Island",
-        description:
-          "<p>Little Saint Simons Island is an island on the Georgia Coast, one of the most protected conservation areas in the state. This 8,840-acre island lies north and northeast of St. Simons Island and is separated from the larger island by the Hampton River. Little St. Simons is only accessible by water and is privately owned.</p><p><br></p><p>In 1760, the island was given to Samuel Ougspourger, a Swiss colonist, as a crown grant. Years later, the island was bought by two brothers named John and James Graham who planned to develop small-scale agriculture. Major Pierce Butler acquired Little St. Simons in 1774 and established an extensive plantation. During WWII, the island was a patrol station for the U.S. Coast Guard who observed German submarines off the cost. Since the 1970s, the island has had a lodge that accommodates paid guests.</p>",
-        uuid: "e0a98bec-543b-47c7-a00b-b3f509dde77f",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-81.394624438855161, 31.304646812090915],
-            [-81.390605697060764, 31.308288903063524],
-            [-81.386817996955912, 31.311725425675213],
-            [-81.383531978984138, 31.313832490634205],
-            [-81.376671357317846, 31.313120536058651],
-            [-81.367808200462136, 31.312181061679169],
-            [-81.363711655888736, 31.312244738537824],
-            [-81.350891381679773, 31.31133203689712],
-            [-81.3389382783039, 31.310975809925022],
-            [-81.336634764026869, 31.31029698405073],
-            [-81.33365943029429, 31.306489010867843],
-            [-81.331782844585433, 31.305739807379112],
-            [-81.3309421491347, 31.305744982954977],
-            [-81.330011453477468, 31.306310724449503],
-            [-81.32823634087751, 31.30731067870887],
-            [-81.324965074944771, 31.308068691477782],
-            [-81.315663634341988, 31.308779721647248],
-            [-81.305825779087328, 31.296923101078942],
-            [-81.295062931228898, 31.291144629707251],
-            [-81.286194538994451, 31.293932365334133],
-            [-81.276435793308323, 31.300732511357815],
-            [-81.271655661593087, 31.279569104770115],
-            [-81.272879766981006, 31.24906847885444],
-            [-81.291937235200649, 31.209045999152917],
-            [-81.311449863684118, 31.221149816963717],
-            [-81.328567707550206, 31.245004821434168],
-            [-81.331967397970715, 31.24679280676644],
-            [-81.332025597022565, 31.248946152703368],
-            [-81.329636458816665, 31.258543437053067],
-            [-81.326828239848652, 31.258387424888177],
-            [-81.324098026963071, 31.25386307210637],
-            [-81.320977783665271, 31.250430804478789],
-            [-81.316531436965917, 31.250040774066566],
-            [-81.313918233204006, 31.252107935251356],
-            [-81.313879230162783, 31.259947546537074],
-            [-81.312143573912806, 31.273177891963702],
-            [-81.315702055237949, 31.276891089868197],
-            [-81.31590834401041, 31.283904908132239],
-            [-81.320498269197913, 31.287050811912437],
-            [-81.325430386626692, 31.290573333252276],
-            [-81.333830485238963, 31.291798347633232],
-            [-81.341964934761947, 31.296598965133182],
-            [-81.340439811989398, 31.298868781357864],
-            [-81.345052649108212, 31.298724630197903],
-            [-81.352157241992089, 31.299342420883455],
-            [-81.354751962871418, 31.300495630163159],
-            [-81.356484035553493, 31.300631571389488],
-            [-81.362230510195474, 31.297910423382699],
-            [-81.364127525948177, 31.295408864148353],
-            [-81.365482537200123, 31.295930022322175],
-            [-81.370102301934082, 31.296045878795415],
-            [-81.37253520156483, 31.300452417267806],
-            [-81.373836746901887, 31.300865102374683],
-            [-81.374789097148522, 31.30019845720204],
-            [-81.375074802222514, 31.298547716774539],
-            [-81.376177197899068, 31.297139475274662],
-            [-81.378396309694793, 31.29529808463565],
-            [-81.380064578136981, 31.295565637121662],
-            [-81.381323648659375, 31.29433804336232],
-            [-81.381674636002387, 31.293018317693718],
-            [-81.379464560296697, 31.290626720130138],
-            [-81.379425201938957, 31.288808364002858],
-            [-81.381435722157278, 31.284065711737348],
-            [-81.38234370022623, 31.284519700771824],
-            [-81.382762790674747, 31.285825000486671],
-            [-81.382259003695765, 31.287052981247953],
-            [-81.38203612423483, 31.289012440213163],
-            [-81.382968695526984, 31.290885045217994],
-            [-81.384323332059282, 31.291965619616391],
-            [-81.385464826760213, 31.292058048337115],
-            [-81.386941402524954, 31.291782248881383],
-            [-81.38738090927086, 31.290760389459273],
-            [-81.386238210314048, 31.289014877329361],
-            [-81.384788470313822, 31.286878896764762],
-            [-81.38456860997826, 31.28599773778285],
-            [-81.385626092496921, 31.285070406958791],
-            [-81.394575762233927, 31.292139233715023],
-            [-81.396318557223111, 31.296665659589735],
-            [-81.396492243018258, 31.299847593060726],
-            [-81.394624438855161, 31.304646812090915],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Sapelo Island",
-        description:
-          "<p>Sapelo Island is home to the Gullah/Geechee people, many of whom reside at Hog Hammock on the southern half of the island. The rest of Sapelo is state-owned and managed. Its ecosystem and long history of occupation make Sapelo a site of great ecological, archaeological, cultural, and historical significance.</p>",
-        uuid: "75fc9ef3-7a0b-4856-83dd-ea8c574eef5f",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-81.237263968060574, 31.444036143671255],
-            [-81.232917649154075, 31.432335396356908],
-            [-81.268696137020086, 31.377176894230139],
-            [-81.2824112240354, 31.37508981577129],
-            [-81.291057691936345, 31.395662446294246],
-            [-81.299304551065774, 31.406505246867312],
-            [-81.311630322459308, 31.426968623177007],
-            [-81.304279841076649, 31.441518082252717],
-            [-81.303312040307276, 31.451832074317174],
-            [-81.305181400950019, 31.453770670539274],
-            [-81.30452856243673, 31.456622020472452],
-            [-81.297394985005468, 31.461309388286843],
-            [-81.298099639163993, 31.463736530388434],
-            [-81.299352357668042, 31.467847012979842],
-            [-81.298287148097529, 31.471038440006922],
-            [-81.296627193589728, 31.474482950309984],
-            [-81.292874735462064, 31.477878031473104],
-            [-81.291108860065066, 31.480772058175003],
-            [-81.294235908504774, 31.483988450855851],
-            [-81.294235908504774, 31.487651564742372],
-            [-81.269562471630749, 31.498082522004871],
-            [-81.237905730769256, 31.530246319827885],
-            [-81.227183048524665, 31.528527002773561],
-            [-81.217912344252298, 31.528527002773561],
-            [-81.216303379048014, 31.525615541927699],
-            [-81.219770406727804, 31.52215769933051],
-            [-81.219734635453662, 31.521552413261052],
-            [-81.219466142087242, 31.5209982060823],
-            [-81.218968473471307, 31.519996781758895],
-            [-81.218407899663774, 31.518754038166477],
-            [-81.218111892875285, 31.518091925382787],
-            [-81.217596859200498, 31.51676346030645],
-            [-81.217058840634195, 31.515376065299225],
-            [-81.216841019829218, 31.514816580436825],
-            [-81.216633058129673, 31.514388139384209],
-            [-81.216304617024875, 31.51380004798601],
-            [-81.215723522042637, 31.513170894352523],
-            [-81.215306067037446, 31.512976969152692],
-            [-81.215046826319494, 31.51285835391884],
-            [-81.214687010350573, 31.512695695253875],
-            [-81.214279087869514, 31.51264945976542],
-            [-81.213481063317559, 31.512566470625003],
-            [-81.213202464253271, 31.512532421909967],
-            [-81.212308209587988, 31.512187265912068],
-            [-81.212197780109562, 31.511742057014285],
-            [-81.212525316547072, 31.511133188857688],
-            [-81.213467840785128, 31.510304903181876],
-            [-81.214475421806881, 31.509405081079727],
-            [-81.214629999778097, 31.509019690930344],
-            [-81.215095592933523, 31.507822681839272],
-            [-81.215224056991914, 31.507204717132719],
-            [-81.215631486116223, 31.506856950141678],
-            [-81.219058782750508, 31.505320061458768],
-            [-81.219560648543492, 31.504632532166557],
-            [-81.219933934306383, 31.504112520273278],
-            [-81.220433577333011, 31.503863217217017],
-            [-81.221797091307479, 31.503592228023933],
-            [-81.22216982598384, 31.503084478999405],
-            [-81.222155522767267, 31.501294219542586],
-            [-81.222368416326375, 31.500761815810893],
-            [-81.222669905806697, 31.500693409545086],
-            [-81.224280425602956, 31.502451856125973],
-            [-81.224615002909204, 31.502469508153784],
-            [-81.225011078388007, 31.50227252262119],
-            [-81.226110331931579, 31.501236346045022],
-            [-81.227031799046671, 31.500134501753859],
-            [-81.22718428422489, 31.499420563338983],
-            [-81.227094372555229, 31.499127731562318],
-            [-81.226785807980235, 31.49895190974464],
-            [-81.22629916412248, 31.498796115565124],
-            [-81.225405876848271, 31.498508849462858],
-            [-81.224558886048072, 31.498241273027215],
-            [-81.223945742959458, 31.497744423253096],
-            [-81.223287117883956, 31.496433100955461],
-            [-81.222880158772199, 31.496103881244458],
-            [-81.222467832203108, 31.495962579413501],
-            [-81.221741643891505, 31.495715780323653],
-            [-81.22049140563405, 31.495291535345302],
-            [-81.219797765666527, 31.495052910617186],
-            [-81.218386170584907, 31.494921589808648],
-            [-81.216578753651476, 31.493685245555586],
-            [-81.209192315584289, 31.487327628209034],
-            [-81.207676620863808, 31.485921353272104],
-            [-81.206950971267901, 31.484288641681317],
-            [-81.207559998607323, 31.483822152655378],
-            [-81.208415228488207, 31.483718488427392],
-            [-81.210890211931385, 31.484133145339339],
-            [-81.21218601478121, 31.484781046764255],
-            [-81.214453871309715, 31.485166276625534],
-            [-81.217095142437259, 31.485869359960844],
-            [-81.218346815089191, 31.48557382614025],
-            [-81.218972651415157, 31.485313061004433],
-            [-81.219752754662096, 31.484550342099844],
-            [-81.220012755090778, 31.483510340385145],
-            [-81.220106614296483, 31.48268510578713],
-            [-81.219892110469331, 31.481420272875308],
-            [-81.220024467775175, 31.478834834019807],
-            [-81.220729104350298, 31.477763786425626],
-            [-81.221771966481469, 31.477200077165531],
-            [-81.223153054168705, 31.475931731330316],
-            [-81.224956923801003, 31.474325159939045],
-            [-81.227916397416507, 31.472521290306741],
-            [-81.228733775843637, 31.471647540953594],
-            [-81.229382041492755, 31.470125525951339],
-            [-81.229209780406578, 31.46779806854012],
-            [-81.227312116822191, 31.465237728783404],
-            [-81.22254881788021, 31.462890444625014],
-            [-81.221455267204945, 31.458503589266432],
-            [-81.222893193141815, 31.455611864852603],
-            [-81.22720741794268, 31.455155064579571],
-            [-81.231470887157656, 31.453683152588688],
-            [-81.234008666452283, 31.449521194545497],
-            [-81.235882368808774, 31.445970382623766],
-            [-81.237263968060574, 31.444036143671255],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Saint Catherines Island",
-        description:
-          "<p>Saint Catherines Island is a sea island, located 42 miles away from Savannah. The island is owned by St Catherines Island Foundation as a private nature preserve and research site, and it is not open to public.</p>",
-        uuid: "643bf294-badc-486b-8644-84209c49735b",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-81.167424857529639, 31.666611913476181],
-            [-81.161571396098623, 31.667881960163626],
-            [-81.154417909190485, 31.687316005764778],
-            [-81.152929012995557, 31.699041063299791],
-            [-81.140557155525087, 31.700917833768553],
-            [-81.127253884816298, 31.694631672884181],
-            [-81.13047006015249, 31.686371949861694],
-            [-81.127244908363934, 31.632159266414003],
-            [-81.15916348644248, 31.569598853380061],
-            [-81.173318439138455, 31.553543260628789],
-            [-81.181357907524543, 31.557880342258127],
-            [-81.18791642120793, 31.570045327316024],
-            [-81.188445333601763, 31.576392276041886],
-            [-81.177297128501849, 31.592621175880659],
-            [-81.179448307583101, 31.601942951899378],
-            [-81.182890194113085, 31.607679429449359],
-            [-81.181312662786837, 31.619008972610565],
-            [-81.185758432888079, 31.626179569548043],
-            [-81.187049140336825, 31.631485811281774],
-            [-81.189632027400236, 31.641986839975978],
-            [-81.188601515634659, 31.647943700669614],
-            [-81.188859659109056, 31.65346219506857],
-            [-81.194213218064931, 31.657343369941572],
-            [-81.197690667089304, 31.659687700946449],
-            [-81.197057941726854, 31.661772169873153],
-            [-81.193852540193333, 31.667053926909102],
-            [-81.185808454002, 31.666429816773569],
-            [-81.173338827588182, 31.669225866597031],
-            [-81.167424857529639, 31.666611913476181],
-          ],
-        ],
-      },
-    },
-    {
-      type: "Feature",
-      properties: {
-        name: "Wassaw Island",
-        description:
-          "<p>One of the islands protected by the Wassaw National Wildlife Refuge, known to support bottle-nose dolphins, loggerhead sea turtles, and many species of birds. Historical structures of interest include a housing compound built by George Parsons, who took ownership of the island in 1866, and Fort Morgan, constructed by local civilians during the Spanish-American War.</p>",
-        uuid: "46c9700b-2dfa-4aed-81e6-50d81c168401",
-      },
-      geometry: {
-        type: "Polygon",
-        coordinates: [
-          [
-            [-81.008573335588196, 31.904005369381931],
-            [-81.011177151790179, 31.907045386452509],
-            [-81.010787451931009, 31.907783363942709],
-            [-81.010690026966202, 31.908336843176514],
-            [-81.010547636633035, 31.908979383858444],
-            [-81.010705015422332, 31.90922113067619],
-            [-81.010907359579974, 31.909335642105031],
-            [-81.011199634474337, 31.909431068186862],
-            [-81.011596828561579, 31.909615558331136],
-            [-81.011911586140158, 31.909850941771278],
-            [-81.012080321930569, 31.910052295329802],
-            [-81.012424470956603, 31.910344435734139],
-            [-81.012721383841807, 31.910476185024777],
-            [-81.013085776928179, 31.910579293033692],
-            [-81.013409681893862, 31.910739683040116],
-            [-81.013625618537645, 31.910900072766953],
-            [-81.013699846758939, 31.911180754116163],
-            [-81.013753830919882, 31.911507259914444],
-            [-81.013787571020472, 31.912039977413698],
-            [-81.013672854678475, 31.913105403158919],
-            [-81.013686857247549, 31.913484571607292],
-            [-81.013601666068752, 31.913614238539751],
-            [-81.013513537263137, 31.913666603979916],
-            [-81.013396032188936, 31.91365662961266],
-            [-81.013340217278696, 31.91374639887902],
-            [-81.013378406427805, 31.914000744658058],
-            [-81.013263831273335, 31.914256656150073],
-            [-81.013177958048615, 31.91433885419999],
-            [-81.012871537397032, 31.91484972623098],
-            [-81.012822499513348, 31.914933825798073],
-            [-81.012792476319277, 31.914975450807884],
-            [-81.012750443847551, 31.915011978862207],
-            [-81.012710412922118, 31.915033216096429],
-            [-81.012659373492184, 31.915048506902032],
-            [-81.012720420653466, 31.915180177622865],
-            [-81.01280648714318, 31.915194618916264],
-            [-81.012850521161184, 31.915220953033678],
-            [-81.012876541262713, 31.915275320219987],
-            [-81.012900398300502, 31.915350037346535],
-            [-81.012914146025608, 31.915396715265299],
-            [-81.012831659675001, 31.915531562453126],
-            [-81.012785833924667, 31.91557953688574],
-            [-81.012736953124318, 31.915611952028737],
-            [-81.012716397790783, 31.915666280421821],
-            [-81.012716397790783, 31.915719441200228],
-            [-81.012701122540662, 31.915793347597226],
-            [-81.012590166437107, 31.91597614729563],
-            [-81.01248340486346, 31.91611207992332],
-            [-81.012352924253378, 31.916221974735141],
-            [-81.012224070593433, 31.916301825748452],
-            [-81.012103134790522, 31.916366201264388],
-            [-81.012008845859469, 31.916383600044739],
-            [-81.011914556928389, 31.916399258944246],
-            [-81.011832566553537, 31.916406218454277],
-            [-81.01172187954748, 31.916406218454277],
-            [-81.011637839413268, 31.916394039311371],
-            [-81.01153740120408, 31.916345322723611],
-            [-81.011465059610558, 31.916280372173606],
-            [-81.011392075264851, 31.916170459608054],
-            [-81.011300256249314, 31.916102513592751],
-            [-81.011215500234968, 31.916088524701017],
-            [-81.01114251588929, 31.916100515179775],
-            [-81.011050696873738, 31.916146478667194],
-            [-81.010996547197905, 31.916208429418276],
-            [-81.01083409817042, 31.916400276640797],
-            [-81.010714027150087, 31.916412267078922],
-            [-81.010549223788871, 31.916416263891282],
-            [-81.010285538410898, 31.916410268672678],
-            [-81.010094837378631, 31.916380292573777],
-            [-81.009904719900817, 31.916301898497878],
-            [-81.009673201669983, 31.916252548771933],
-            [-81.009468676596086, 31.916207605248488],
-            [-81.009309832249357, 31.916168830426319],
-            [-81.009201859621513, 31.916096568214098],
-            [-81.009114650960569, 31.916055149603444],
-            [-81.009053397258214, 31.916051624614447],
-            [-81.008999410944313, 31.916058674592318],
-            [-81.008951653820446, 31.916071893299357],
-            [-81.008926737060193, 31.916096568214098],
-            [-81.008926283864383, 31.9161408749397],
-            [-81.008952762189779, 31.916230851305869],
-            [-81.008926283864383, 31.91632022946327],
-            [-81.008938713168277, 31.916419401814409],
-            [-81.00897102935842, 31.916569215163374],
-            [-81.009035661738722, 31.916691597718085],
-            [-81.009025280548187, 31.916797880420077],
-            [-81.008989512443591, 31.91694534590513],
-            [-81.008910311640577, 31.917294491184649],
-            [-81.008815781649886, 31.917621949217665],
-            [-81.00873896132245, 31.917794954663343],
-            [-81.008607784662331, 31.91820996219802],
-            [-81.008580729900558, 31.918929711696354],
-            [-81.008623483821239, 31.919352672254792],
-            [-81.00871719211554, 31.919770250837804],
-            [-81.008709914164896, 31.920141608266118],
-            [-81.008830429630962, 31.920584938938777],
-            [-81.008837673812522, 31.921341229534598],
-            [-81.008656418121092, 31.922277488498722],
-            [-81.008446620735455, 31.924903340892413],
-            [-81.008484466610412, 31.92610788478051],
-            [-81.008886066480741, 31.928609760804932],
-            [-81.004575518190066, 31.931253697587632],
-            [-81.000443684633424, 31.929628494591078],
-            [-80.99014921604757, 31.930531867391139],
-            [-80.977227732506364, 31.926018198482904],
-            [-80.967580871506414, 31.918583920281108],
-            [-80.929717940531361, 31.910538326755773],
-            [-80.988403454125915, 31.855680898900978],
-            [-81.000812562080853, 31.855418272277593],
-            [-81.004620648119925, 31.860145551498519],
-            [-80.997726699256077, 31.869731423252063],
-            [-81.001439592835538, 31.880548715467302],
-            [-81.006164108808832, 31.887230930513873],
-            [-81.005970081675869, 31.89162079439696],
-            [-81.008573335588196, 31.904005369381931],
-          ],
-        ],
-      },
-    },
-  ],
-};
