+++ package.json
@@ -56,7 +56,7 @@
     "eslint-plugin-react": "^7.34.2",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-unused-imports": "^4.1.4",
-    "nuka-carousel": "^8.0.1",
+    "nuka-carousel": "^8.2.0",
     "postcss": "^8.4.38",
     "prettier": "^3.2.5",
     "react-instantsearch": "^7.11.4",
